Module lp55s69_pac::adc0::cmdh15

ADC Command High Buffer Register

Structs

R
Value read from the register

W
Value to write to the register

_AVGSW
Proxy

_LOOPW
Proxy

_LWIW
Proxy

_NEXTW
Proxy

_STSW
Proxy

_WAIT_TRIGW
Proxy

Enums

AVGSR
Possible values of the field AVGS

AVGSW
Values that can be written to the field AVGS

LOOPR
Possible values of the field LOOP

LOOPW
Values that can be written to the field LOOP

LWIR
Possible values of the field LWI

LWIW
Values that can be written to the field LWI

NEXTR
Possible values of the field NEXT

NEXTW
Values that can be written to the field NEXT

STSR
Possible values of the field STS

STSW
Values that can be written to the field STS

WAIT_TRIGR
Possible values of the field WAIT_TRIG

WAIT_TRIGW
Values that can be written to the field WAIT_TRIG