Module lp55s69_pac::adc0::ie

Interrupt Enable Register

Structs

R
Value read from the register

W
Value to write to the register

_FOFIE0W
Proxy

_FOFIE1W
Proxy

_FWMIE0W
Proxy

_FWMIE1W
Proxy

_TCOMP_IEW
Proxy

_TEXC_IEW
Proxy

Enums

FOFIE0R
Possible values of the field FOFIE0

FOFIE0W
Values that can be written to the field FOFIE0

FOFIE1R
Possible values of the field FOFIE1

FOFIE1W
Values that can be written to the field FOFIE1

FWMIE0R
Possible values of the field FWMIE0

FWMIE0W
Values that can be written to the field FWMIE0

FWMIE1R
Possible values of the field FWMIE1

FWMIE1W
Values that can be written to the field FWMIE1

TCOMP_IER
Possible values of the field TCOMP_IE

TCOMP_IEW
Values that can be written to the field TCOMP_IE

TEXC_IER
Possible values of the field TEXC_IE

TEXC_IEW
Values that can be written to the field TEXC_IE