Module lp55s69_pac::adc0::stat

ADC Status Register

Structs

R
Value read from the register

W
Value to write to the register

_FOF0W
Proxy

_FOF1W
Proxy

_TCOMP_INTW
Proxy

_TEXC_INTW
Proxy

Enums

ADC_ACTIVER
Possible values of the field ADC_ACTIVE

CAL_RDYR
Possible values of the field CAL_RDY

CMDACTR
Possible values of the field CMDACT

FOF0R
Possible values of the field FOF0

FOF0W
Values that can be written to the field FOF0

FOF1R
Possible values of the field FOF1

FOF1W
Values that can be written to the field FOF1

RDY0R
Possible values of the field RDY0

RDY1R
Possible values of the field RDY1

TCOMP_INTR
Possible values of the field TCOMP_INT

TCOMP_INTW
Values that can be written to the field TCOMP_INT

TEXC_INTR
Possible values of the field TEXC_INT

TEXC_INTW
Values that can be written to the field TEXC_INT

TRGACTR
Possible values of the field TRGACT