Module lp55s69_pac::ahb_secure_ctrl::master_sec_level

master secure level register

Structs

R
Value read from the register

W
Value to write to the register

_HASHW
Proxy

_MASTER_SEC_LEVEL_LOCKW
Proxy

_MCM33CW
Proxy

_MCM33SW
Proxy

_PQW
Proxy

_SDIOW
Proxy

_SDMA0W
Proxy

_SDMA1W
Proxy

_USBFSDW
Proxy

_USBFSHW
Proxy

Enums

HASHR
Possible values of the field HASH

HASHW
Values that can be written to the field HASH

MASTER_SEC_LEVEL_LOCKR
Possible values of the field MASTER_SEC_LEVEL_LOCK

MASTER_SEC_LEVEL_LOCKW
Values that can be written to the field MASTER_SEC_LEVEL_LOCK

MCM33CR
Possible values of the field MCM33C

MCM33SR
Possible values of the field MCM33S

MCM33CW
Values that can be written to the field MCM33C

MCM33SW
Values that can be written to the field MCM33S

PQR
Possible values of the field PQ

PQW
Values that can be written to the field PQ

SDIOR
Possible values of the field SDIO

SDIOW
Values that can be written to the field SDIO

SDMA0R
Possible values of the field SDMA0

SDMA0W
Values that can be written to the field SDMA0

SDMA1R
Possible values of the field SDMA1

SDMA1W
Values that can be written to the field SDMA1

USBFSDR
Possible values of the field USBFSD

USBFSDW
Values that can be written to the field USBFSD

USBFSHR
Possible values of the field USBFSH

USBFSHW
Values that can be written to the field USBFSH