Struct lp55s69_pac::ahb_secure_ctrl::sec_cpu_int_mask0::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn sys_irq(&mut self) -> _SYS_IRQW

Bit 0 - Watchdog Timer, Brown Out Detectors and Flash Controller interrupts

pub fn sdma0_irq(&mut self) -> _SDMA0_IRQW

Bit 1 - System DMA 0 (non-secure) interrupt.

pub fn gpio_globalint0_irq(&mut self) -> _GPIO_GLOBALINT0_IRQW

Bit 2 - GPIO Group 0 interrupt.

pub fn gpio_globalint1_irq(&mut self) -> _GPIO_GLOBALINT1_IRQW

Bit 3 - GPIO Group 1 interrupt.

pub fn gpio_int0_irq0(&mut self) -> _GPIO_INT0_IRQ0W

Bit 4 - Pin interrupt 0 or pattern match engine slice 0 interrupt.

pub fn gpio_int0_irq1(&mut self) -> _GPIO_INT0_IRQ1W

Bit 5 - Pin interrupt 1 or pattern match engine slice 1 interrupt.

pub fn gpio_int0_irq2(&mut self) -> _GPIO_INT0_IRQ2W

Bit 6 - Pin interrupt 2 or pattern match engine slice 2 interrupt.

pub fn gpio_int0_irq3(&mut self) -> _GPIO_INT0_IRQ3W

Bit 7 - Pin interrupt 3 or pattern match engine slice 3 interrupt.

pub fn utick_irq(&mut self) -> _UTICK_IRQW

Bit 8 - Micro Tick Timer interrupt.

pub fn mrt_irq(&mut self) -> _MRT_IRQW

Bit 9 - Multi-Rate Timer interrupt.

pub fn ctimer0_irq(&mut self) -> _CTIMER0_IRQW

Bit 10 - Standard counter/timer 0 interrupt.

pub fn ctimer1_irq(&mut self) -> _CTIMER1_IRQW

Bit 11 - Standard counter/timer 1 interrupt.

pub fn sct_irq(&mut self) -> _SCT_IRQW

Bit 12 - SCTimer/PWM interrupt.

pub fn ctimer3_irq(&mut self) -> _CTIMER3_IRQW

Bit 13 - Standard counter/timer 3 interrupt.

pub fn flexcomm0_irq(&mut self) -> _FLEXCOMM0_IRQW

Bit 14 - Flexcomm 0 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm1_irq(&mut self) -> _FLEXCOMM1_IRQW

Bit 15 - Flexcomm 1 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm2_irq(&mut self) -> _FLEXCOMM2_IRQW

Bit 16 - Flexcomm 2 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm3_irq(&mut self) -> _FLEXCOMM3_IRQW

Bit 17 - Flexcomm 3 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm4_irq(&mut self) -> _FLEXCOMM4_IRQW

Bit 18 - Flexcomm 4 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm5_irq(&mut self) -> _FLEXCOMM5_IRQW

Bit 19 - Flexcomm 5 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm6_irq(&mut self) -> _FLEXCOMM6_IRQW

Bit 20 - Flexcomm 6 interrupt (USART, SPI, I2C, I2S).

pub fn flexcomm7_irq(&mut self) -> _FLEXCOMM7_IRQW

Bit 21 - Flexcomm 7 interrupt (USART, SPI, I2C, I2S).

pub fn adc_irq(&mut self) -> _ADC_IRQW

Bit 22 - General Purpose ADC interrupt.

pub fn reserved0(&mut self) -> _RESERVED0W

Bit 23 - Reserved. Read value is undefined, only zero should be written.

pub fn acmp_capt0_irq(&mut self) -> _ACMP_CAPT0_IRQW

Bit 24 - Analog Comparator interrupt.

pub fn reserved1(&mut self) -> _RESERVED1W

Bit 25 - Reserved. Read value is undefined, only zero should be written.

pub fn reserved2(&mut self) -> _RESERVED2W

Bit 26 - Reserved. Read value is undefined, only zero should be written.

pub fn usb0_needclk(&mut self) -> _USB0_NEEDCLKW

Bit 27 - USB Full Speed Controller Clock request interrupt.

pub fn usb0_irq(&mut self) -> _USB0_IRQW

Bit 28 - USB High Speed Controller interrupt.

pub fn rtc_irq(&mut self) -> _RTC_IRQW

Bit 29 - RTC_LITE0_ALARM_IRQ, RTC_LITE0_WAKEUP_IRQ

pub fn reserved3(&mut self) -> _RESERVED3W

Bit 30 - Reserved. Read value is undefined, only zero should be written.

pub fn mailbox_irq(&mut self) -> _MAILBOX_IRQW

Bit 31 - Mailbox interrupt.