Module lp55s69_pac::ahb_secure_ctrl::sec_cpu_int_mask1

Secure Interrupt mask for CPU1

Structs

R
Value read from the register

W
Value to write to the register

_CASPER_IRQW
Proxy

_CTIMER2_IRQW
Proxy

_CTIMER4_IRQW
Proxy

_GPIO_INT0_IRQ4W
Proxy

_GPIO_INT0_IRQ5W
Proxy

_GPIO_INT0_IRQ6W
Proxy

_GPIO_INT0_IRQ7W
Proxy

_LSPI_HS_IRQW
Proxy

_OS_EVENT_TIMER_IRQW
Proxy

_PLU_IRQW
Proxy

_PQ_IRQW
Proxy

_QDDKEY_IRQW
Proxy

_RESERVED0W
Proxy

_RESERVED1W
Proxy

_RESERVED2W
Proxy

_RESERVED3W
Proxy

_RESERVED4W
Proxy

_RESERVED5W
Proxy

_SDIO_IRQW
Proxy

_SDMA1_IRQW
Proxy

_SEC_GPIO_INT0_IRQ0W
Proxy

_SEC_GPIO_INT0_IRQ1W
Proxy

_SEC_HYPERVISOR_CALL_IRQW
Proxy

_SEC_VIO_IRQW
Proxy

_SHA_IRQW
Proxy

_USB1_UTMI_IRQW
Proxy

_USB1_IRQW
Proxy

_USB1_NEEDCLKW
Proxy

Enums

CASPER_IRQR
Possible values of the field CASPER_IRQ

CASPER_IRQW
Values that can be written to the field CASPER_IRQ

CTIMER2_IRQR
Possible values of the field CTIMER2_IRQ

CTIMER2_IRQW
Values that can be written to the field CTIMER2_IRQ

CTIMER4_IRQR
Possible values of the field CTIMER4_IRQ

CTIMER4_IRQW
Values that can be written to the field CTIMER4_IRQ

GPIO_INT0_IRQ4R
Possible values of the field GPIO_INT0_IRQ4

GPIO_INT0_IRQ4W
Values that can be written to the field GPIO_INT0_IRQ4

GPIO_INT0_IRQ5R
Possible values of the field GPIO_INT0_IRQ5

GPIO_INT0_IRQ5W
Values that can be written to the field GPIO_INT0_IRQ5

GPIO_INT0_IRQ6R
Possible values of the field GPIO_INT0_IRQ6

GPIO_INT0_IRQ6W
Values that can be written to the field GPIO_INT0_IRQ6

GPIO_INT0_IRQ7R
Possible values of the field GPIO_INT0_IRQ7

GPIO_INT0_IRQ7W
Values that can be written to the field GPIO_INT0_IRQ7

LSPI_HS_IRQR
Possible values of the field LSPI_HS_IRQ

LSPI_HS_IRQW
Values that can be written to the field LSPI_HS_IRQ

OS_EVENT_TIMER_IRQR
Possible values of the field OS_EVENT_TIMER_IRQ

OS_EVENT_TIMER_IRQW
Values that can be written to the field OS_EVENT_TIMER_IRQ

PLU_IRQR
Possible values of the field PLU_IRQ

PLU_IRQW
Values that can be written to the field PLU_IRQ

PQ_IRQR
Possible values of the field PQ_IRQ

PQ_IRQW
Values that can be written to the field PQ_IRQ

QDDKEY_IRQR
Possible values of the field QDDKEY_IRQ

QDDKEY_IRQW
Values that can be written to the field QDDKEY_IRQ

RESERVED0R
Possible values of the field RESERVED0

RESERVED0W
Values that can be written to the field RESERVED0

RESERVED1R
Possible values of the field RESERVED1

RESERVED1W
Values that can be written to the field RESERVED1

RESERVED2R
Possible values of the field RESERVED2

RESERVED2W
Values that can be written to the field RESERVED2

RESERVED3R
Possible values of the field RESERVED3

RESERVED3W
Values that can be written to the field RESERVED3

RESERVED4R
Possible values of the field RESERVED4

RESERVED4W
Values that can be written to the field RESERVED4

RESERVED5R
Possible values of the field RESERVED5

RESERVED5W
Values that can be written to the field RESERVED5

SDIO_IRQR
Possible values of the field SDIO_IRQ

SDIO_IRQW
Values that can be written to the field SDIO_IRQ

SDMA1_IRQR
Possible values of the field SDMA1_IRQ

SDMA1_IRQW
Values that can be written to the field SDMA1_IRQ

SEC_GPIO_INT0_IRQ0R
Possible values of the field SEC_GPIO_INT0_IRQ0

SEC_GPIO_INT0_IRQ0W
Values that can be written to the field SEC_GPIO_INT0_IRQ0

SEC_GPIO_INT0_IRQ1R
Possible values of the field SEC_GPIO_INT0_IRQ1

SEC_GPIO_INT0_IRQ1W
Values that can be written to the field SEC_GPIO_INT0_IRQ1

SEC_HYPERVISOR_CALL_IRQR
Possible values of the field SEC_HYPERVISOR_CALL_IRQ

SEC_HYPERVISOR_CALL_IRQW
Values that can be written to the field SEC_HYPERVISOR_CALL_IRQ

SEC_VIO_IRQR
Possible values of the field SEC_VIO_IRQ

SEC_VIO_IRQW
Values that can be written to the field SEC_VIO_IRQ

SHA_IRQR
Possible values of the field SHA_IRQ

SHA_IRQW
Values that can be written to the field SHA_IRQ

USB1_UTMI_IRQR
Possible values of the field USB1_UTMI_IRQ

USB1_IRQR
Possible values of the field USB1_IRQ

USB1_NEEDCLKR
Possible values of the field USB1_NEEDCLK

USB1_UTMI_IRQW
Values that can be written to the field USB1_UTMI_IRQ

USB1_IRQW
Values that can be written to the field USB1_IRQ

USB1_NEEDCLKW
Values that can be written to the field USB1_NEEDCLK