Module lp55s69_pac::ahb_secure_ctrl::sec_ctrl_apb_bridge0_mem_ctrl0

Security access rules for APB Bridge 0 peripherals. Each APB bridge sector is 4 Kbytes. There are 32 APB Bridge 0 sectors in total.

Structs

R
Value read from the register

W
Value to write to the register

_GINT0_RULEW
Proxy

_GINT1_RULEW
Proxy

_IOCON_RULEW
Proxy

_PINT_RULEW
Proxy

_PMUX_RULEW
Proxy

_SEC_PINT_RULEW
Proxy

_SYSCON_RULEW
Proxy

Enums

GINT0_RULER
Possible values of the field GINT0_RULE

GINT0_RULEW
Values that can be written to the field GINT0_RULE

GINT1_RULER
Possible values of the field GINT1_RULE

GINT1_RULEW
Values that can be written to the field GINT1_RULE

IOCON_RULER
Possible values of the field IOCON_RULE

IOCON_RULEW
Values that can be written to the field IOCON_RULE

PINT_RULER
Possible values of the field PINT_RULE

PINT_RULEW
Values that can be written to the field PINT_RULE

PMUX_RULER
Possible values of the field PMUX_RULE

PMUX_RULEW
Values that can be written to the field PMUX_RULE

SEC_PINT_RULER
Possible values of the field SEC_PINT_RULE

SEC_PINT_RULEW
Values that can be written to the field SEC_PINT_RULE

SYSCON_RULER
Possible values of the field SYSCON_RULE

SYSCON_RULEW
Values that can be written to the field SYSCON_RULE