Module lp55s69_pac::ahb_secure_ctrl::sec_gpio_mask1

Secure GPIO mask for port 1 pins.

Structs

R
Value read from the register

W
Value to write to the register

_PIO1_PIN0_SEC_MASKW
Proxy

_PIO1_PIN1_SEC_MASKW
Proxy

_PIO1_PIN2_SEC_MASKW
Proxy

_PIO1_PIN3_SEC_MASKW
Proxy

_PIO1_PIN4_SEC_MASKW
Proxy

_PIO1_PIN5_SEC_MASKW
Proxy

_PIO1_PIN6_SEC_MASKW
Proxy

_PIO1_PIN7_SEC_MASKW
Proxy

_PIO1_PIN8_SEC_MASKW
Proxy

_PIO1_PIN9_SEC_MASKW
Proxy

_PIO1_PIN10_SEC_MASKW
Proxy

_PIO1_PIN11_SEC_MASKW
Proxy

_PIO1_PIN12_SEC_MASKW
Proxy

_PIO1_PIN13_SEC_MASKW
Proxy

_PIO1_PIN14_SEC_MASKW
Proxy

_PIO1_PIN15_SEC_MASKW
Proxy

_PIO1_PIN16_SEC_MASKW
Proxy

_PIO1_PIN17_SEC_MASKW
Proxy

_PIO1_PIN18_SEC_MASKW
Proxy

_PIO1_PIN19_SEC_MASKW
Proxy

_PIO1_PIN20_SEC_MASKW
Proxy

_PIO1_PIN21_SEC_MASKW
Proxy

_PIO1_PIN22_SEC_MASKW
Proxy

_PIO1_PIN23_SEC_MASKW
Proxy

_PIO1_PIN24_SEC_MASKW
Proxy

_PIO1_PIN25_SEC_MASKW
Proxy

_PIO1_PIN26_SEC_MASKW
Proxy

_PIO1_PIN27_SEC_MASKW
Proxy

_PIO1_PIN28_SEC_MASKW
Proxy

_PIO1_PIN29_SEC_MASKW
Proxy

_PIO1_PIN30_SEC_MASKW
Proxy

_PIO1_PIN31_SEC_MASKW
Proxy

Enums

PIO1_PIN0_SEC_MASKR
Possible values of the field PIO1_PIN0_SEC_MASK

PIO1_PIN0_SEC_MASKW
Values that can be written to the field PIO1_PIN0_SEC_MASK

PIO1_PIN1_SEC_MASKR
Possible values of the field PIO1_PIN1_SEC_MASK

PIO1_PIN1_SEC_MASKW
Values that can be written to the field PIO1_PIN1_SEC_MASK

PIO1_PIN2_SEC_MASKR
Possible values of the field PIO1_PIN2_SEC_MASK

PIO1_PIN2_SEC_MASKW
Values that can be written to the field PIO1_PIN2_SEC_MASK

PIO1_PIN3_SEC_MASKR
Possible values of the field PIO1_PIN3_SEC_MASK

PIO1_PIN3_SEC_MASKW
Values that can be written to the field PIO1_PIN3_SEC_MASK

PIO1_PIN4_SEC_MASKR
Possible values of the field PIO1_PIN4_SEC_MASK

PIO1_PIN4_SEC_MASKW
Values that can be written to the field PIO1_PIN4_SEC_MASK

PIO1_PIN5_SEC_MASKR
Possible values of the field PIO1_PIN5_SEC_MASK

PIO1_PIN5_SEC_MASKW
Values that can be written to the field PIO1_PIN5_SEC_MASK

PIO1_PIN6_SEC_MASKR
Possible values of the field PIO1_PIN6_SEC_MASK

PIO1_PIN6_SEC_MASKW
Values that can be written to the field PIO1_PIN6_SEC_MASK

PIO1_PIN7_SEC_MASKR
Possible values of the field PIO1_PIN7_SEC_MASK

PIO1_PIN7_SEC_MASKW
Values that can be written to the field PIO1_PIN7_SEC_MASK

PIO1_PIN8_SEC_MASKR
Possible values of the field PIO1_PIN8_SEC_MASK

PIO1_PIN8_SEC_MASKW
Values that can be written to the field PIO1_PIN8_SEC_MASK

PIO1_PIN9_SEC_MASKR
Possible values of the field PIO1_PIN9_SEC_MASK

PIO1_PIN9_SEC_MASKW
Values that can be written to the field PIO1_PIN9_SEC_MASK

PIO1_PIN10_SEC_MASKR
Possible values of the field PIO1_PIN10_SEC_MASK

PIO1_PIN10_SEC_MASKW
Values that can be written to the field PIO1_PIN10_SEC_MASK

PIO1_PIN11_SEC_MASKR
Possible values of the field PIO1_PIN11_SEC_MASK

PIO1_PIN11_SEC_MASKW
Values that can be written to the field PIO1_PIN11_SEC_MASK

PIO1_PIN12_SEC_MASKR
Possible values of the field PIO1_PIN12_SEC_MASK

PIO1_PIN12_SEC_MASKW
Values that can be written to the field PIO1_PIN12_SEC_MASK

PIO1_PIN13_SEC_MASKR
Possible values of the field PIO1_PIN13_SEC_MASK

PIO1_PIN13_SEC_MASKW
Values that can be written to the field PIO1_PIN13_SEC_MASK

PIO1_PIN14_SEC_MASKR
Possible values of the field PIO1_PIN14_SEC_MASK

PIO1_PIN14_SEC_MASKW
Values that can be written to the field PIO1_PIN14_SEC_MASK

PIO1_PIN15_SEC_MASKR
Possible values of the field PIO1_PIN15_SEC_MASK

PIO1_PIN15_SEC_MASKW
Values that can be written to the field PIO1_PIN15_SEC_MASK

PIO1_PIN16_SEC_MASKR
Possible values of the field PIO1_PIN16_SEC_MASK

PIO1_PIN16_SEC_MASKW
Values that can be written to the field PIO1_PIN16_SEC_MASK

PIO1_PIN17_SEC_MASKR
Possible values of the field PIO1_PIN17_SEC_MASK

PIO1_PIN17_SEC_MASKW
Values that can be written to the field PIO1_PIN17_SEC_MASK

PIO1_PIN18_SEC_MASKR
Possible values of the field PIO1_PIN18_SEC_MASK

PIO1_PIN18_SEC_MASKW
Values that can be written to the field PIO1_PIN18_SEC_MASK

PIO1_PIN19_SEC_MASKR
Possible values of the field PIO1_PIN19_SEC_MASK

PIO1_PIN19_SEC_MASKW
Values that can be written to the field PIO1_PIN19_SEC_MASK

PIO1_PIN20_SEC_MASKR
Possible values of the field PIO1_PIN20_SEC_MASK

PIO1_PIN20_SEC_MASKW
Values that can be written to the field PIO1_PIN20_SEC_MASK

PIO1_PIN21_SEC_MASKR
Possible values of the field PIO1_PIN21_SEC_MASK

PIO1_PIN21_SEC_MASKW
Values that can be written to the field PIO1_PIN21_SEC_MASK

PIO1_PIN22_SEC_MASKR
Possible values of the field PIO1_PIN22_SEC_MASK

PIO1_PIN22_SEC_MASKW
Values that can be written to the field PIO1_PIN22_SEC_MASK

PIO1_PIN23_SEC_MASKR
Possible values of the field PIO1_PIN23_SEC_MASK

PIO1_PIN23_SEC_MASKW
Values that can be written to the field PIO1_PIN23_SEC_MASK

PIO1_PIN24_SEC_MASKR
Possible values of the field PIO1_PIN24_SEC_MASK

PIO1_PIN24_SEC_MASKW
Values that can be written to the field PIO1_PIN24_SEC_MASK

PIO1_PIN25_SEC_MASKR
Possible values of the field PIO1_PIN25_SEC_MASK

PIO1_PIN25_SEC_MASKW
Values that can be written to the field PIO1_PIN25_SEC_MASK

PIO1_PIN26_SEC_MASKR
Possible values of the field PIO1_PIN26_SEC_MASK

PIO1_PIN26_SEC_MASKW
Values that can be written to the field PIO1_PIN26_SEC_MASK

PIO1_PIN27_SEC_MASKR
Possible values of the field PIO1_PIN27_SEC_MASK

PIO1_PIN27_SEC_MASKW
Values that can be written to the field PIO1_PIN27_SEC_MASK

PIO1_PIN28_SEC_MASKR
Possible values of the field PIO1_PIN28_SEC_MASK

PIO1_PIN28_SEC_MASKW
Values that can be written to the field PIO1_PIN28_SEC_MASK

PIO1_PIN29_SEC_MASKR
Possible values of the field PIO1_PIN29_SEC_MASK

PIO1_PIN29_SEC_MASKW
Values that can be written to the field PIO1_PIN29_SEC_MASK

PIO1_PIN30_SEC_MASKR
Possible values of the field PIO1_PIN30_SEC_MASK

PIO1_PIN30_SEC_MASKW
Values that can be written to the field PIO1_PIN30_SEC_MASK

PIO1_PIN31_SEC_MASKR
Possible values of the field PIO1_PIN31_SEC_MASK

PIO1_PIN31_SEC_MASKW
Values that can be written to the field PIO1_PIN31_SEC_MASK