Module lp55s69_pac::anactrl::ringo0_ctrl

First Ring Oscillator module control register.

Structs

DIVISORR
Value of the field

DIV_UPDATE_REQR
Value of the field

R
Value read from the register

W
Value to write to the register

_DIVISORW
Proxy

_E_IV0W
Proxy

_E_IV1W
Proxy

_E_ND0W
Proxy

_E_ND1W
Proxy

_E_NR0W
Proxy

_E_NR1W
Proxy

_E_PN0W
Proxy

_E_PN1W
Proxy

_FSW
Proxy

_PDW
Proxy

_SLW
Proxy

_SWN_SWPW
Proxy

Enums

E_IV0R
Possible values of the field E_IV0

E_IV0W
Values that can be written to the field E_IV0

E_IV1R
Possible values of the field E_IV1

E_IV1W
Values that can be written to the field E_IV1

E_ND0R
Possible values of the field E_ND0

E_ND0W
Values that can be written to the field E_ND0

E_ND1R
Possible values of the field E_ND1

E_ND1W
Values that can be written to the field E_ND1

E_NR0R
Possible values of the field E_NR0

E_NR0W
Values that can be written to the field E_NR0

E_NR1R
Possible values of the field E_NR1

E_NR1W
Values that can be written to the field E_NR1

E_PN0R
Possible values of the field E_PN0

E_PN0W
Values that can be written to the field E_PN0

E_PN1R
Possible values of the field E_PN1

E_PN1W
Values that can be written to the field E_PN1

FSR
Possible values of the field FS

FSW
Values that can be written to the field FS

PDR
Possible values of the field PD

PDW
Values that can be written to the field PD

SLR
Possible values of the field SL

SLW
Values that can be written to the field SL

SWN_SWPR
Possible values of the field SWN_SWP

SWN_SWPW
Values that can be written to the field SWN_SWP