Module lp55s69_pac::anactrl::ringo1_ctrl

Second Ring Oscillator module control register.

Structs

DIVISORR
Value of the field

DIV_UPDATE_REQR
Value of the field

R
Value read from the register

W
Value to write to the register

_DIVISORW
Proxy

_E_M2W
Proxy

_E_M3W
Proxy

_E_M4W
Proxy

_E_M5W
Proxy

_E_R24W
Proxy

_E_R35W
Proxy

_FSW
Proxy

_PDW
Proxy

_SW
Proxy

Enums

E_M2R
Possible values of the field E_M2

E_M2W
Values that can be written to the field E_M2

E_M3R
Possible values of the field E_M3

E_M3W
Values that can be written to the field E_M3

E_M4R
Possible values of the field E_M4

E_M4W
Values that can be written to the field E_M4

E_M5R
Possible values of the field E_M5

E_M5W
Values that can be written to the field E_M5

E_R24R
Possible values of the field E_R24

E_R24W
Values that can be written to the field E_R24

E_R35R
Possible values of the field E_R35

E_R35W
Values that can be written to the field E_R35

FSR
Possible values of the field FS

FSW
Values that can be written to the field FS

PDR
Possible values of the field PD

PDW
Values that can be written to the field PD

SR
Possible values of the field S

SW
Values that can be written to the field S