Module lp55s69_pac::anactrl::xo32m_ctrl

32 MHz Crystal Oscillator Control register

Structs

AMPR
Value of the field

CAPTESTSTARTR
Value of the field

GMR
Value of the field

OSC_CAP_INR
Value of the field

OSC_CAP_OUTR
Value of the field

R
Value read from the register

SLAVER
Value of the field

W
Value to write to the register

_ACBUF_PASS_ENABLEW
Proxy

_AMPW
Proxy

_CAPTESTENABLEW
Proxy

_CAPTESTOSCINSELW
Proxy

_CAPTESTSTARTSRCSELW
Proxy

_CAPTESTSTARTW
Proxy

_ENABLE_PLL_USB_OUTW
Proxy

_ENABLE_SYSTEM_CLK_OUTW
Proxy

_GMW
Proxy

_OSC_CAP_INW
Proxy

_OSC_CAP_OUTW
Proxy

_SLAVEW
Proxy

Enums

ACBUF_PASS_ENABLER
Possible values of the field ACBUF_PASS_ENABLE

ACBUF_PASS_ENABLEW
Values that can be written to the field ACBUF_PASS_ENABLE

CAPTESTENABLER
Possible values of the field CAPTESTENABLE

CAPTESTENABLEW
Values that can be written to the field CAPTESTENABLE

CAPTESTOSCINSELR
Possible values of the field CAPTESTOSCINSEL

CAPTESTOSCINSELW
Values that can be written to the field CAPTESTOSCINSEL

CAPTESTSTARTSRCSELR
Possible values of the field CAPTESTSTARTSRCSEL

CAPTESTSTARTSRCSELW
Values that can be written to the field CAPTESTSTARTSRCSEL

ENABLE_PLL_USB_OUTR
Possible values of the field ENABLE_PLL_USB_OUT

ENABLE_PLL_USB_OUTW
Values that can be written to the field ENABLE_PLL_USB_OUT

ENABLE_SYSTEM_CLK_OUTR
Possible values of the field ENABLE_SYSTEM_CLK_OUT

ENABLE_SYSTEM_CLK_OUTW
Values that can be written to the field ENABLE_SYSTEM_CLK_OUT