Struct lp55s69_pac::casper::RegisterBlock

#[repr(C)]

pub struct RegisterBlock {
    pub ctrl0: CTRL0,
    pub ctrl1: CTRL1,
    pub loader: LOADER,
    pub status: STATUS,
    pub intenset: INTENSET,
    pub intenclr: INTENCLR,
    pub intstat: INTSTAT,
    pub areg: AREG,
    pub breg: BREG,
    pub creg: CREG,
    pub dreg: DREG,
    pub res0: RES0,
    pub res1: RES1,
    pub res2: RES2,
    pub res3: RES3,
    pub mask: MASK,
    pub remask: REMASK,
    pub lock: LOCK,
    // some fields omitted
}

Register block

Fields
ctrl0: CTRL0

0x00 - Contains the offsets of AB and CD in the RAM.

ctrl1: CTRL1

0x04 - Contains the opcode mode, iteration count, and result offset (in RAM) and also launches the accelerator. Note: with CP version: CTRL0 and CRTL1 can be written in one go with MCRR.

loader: LOADER

0x08 - Contains an optional loader to load into CTRL0/1 in steps to perform a set of operations.

status: STATUS

0x0c - Indicates operational status and would contain the carry bit if used.

intenset: INTENSET

0x10 - Sets interrupts

intenclr: INTENCLR

0x14 - Clears interrupts

intstat: INTSTAT

0x18 - Interrupt status bits (mask of INTENSET and STATUS)

areg: AREG

0x20 - A register

breg: BREG

0x24 - B register

creg: CREG

0x28 - C register

dreg: DREG

0x2c - D register

res0: RES0

0x30 - Result register 0

res1: RES1

0x34 - Result register 1

res2: RES2

0x38 - Result register 2

res3: RES3

0x3c - Result register 3

mask: MASK

0x60 - Optional mask register

remask: REMASK

0x64 - Optional re-mask register

lock: LOCK

0x80 - Security lock register