Struct lp55s69_pac::ctimer0::emr::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn em0(&self) -> EM0R

Bit 0 - External Match 0. This bit reflects the state of output MAT0, whether or not this output is connected to a pin. When a match occurs between the TC and MR0, this bit can either toggle, go LOW, go HIGH, or do nothing, as selected by EMR[5:4]. This bit is driven to the MAT pins if the match function is selected via IOCON. 0 = LOW. 1 = HIGH.

pub fn em1(&self) -> EM1R

Bit 1 - External Match 1. This bit reflects the state of output MAT1, whether or not this output is connected to a pin. When a match occurs between the TC and MR1, this bit can either toggle, go LOW, go HIGH, or do nothing, as selected by EMR[7:6]. This bit is driven to the MAT pins if the match function is selected via IOCON. 0 = LOW. 1 = HIGH.

pub fn em2(&self) -> EM2R

Bit 2 - External Match 2. This bit reflects the state of output MAT2, whether or not this output is connected to a pin. When a match occurs between the TC and MR2, this bit can either toggle, go LOW, go HIGH, or do nothing, as selected by EMR[9:8]. This bit is driven to the MAT pins if the match function is selected via IOCON. 0 = LOW. 1 = HIGH.

pub fn em3(&self) -> EM3R

Bit 3 - External Match 3. This bit reflects the state of output MAT3, whether or not this output is connected to a pin. When a match occurs between the TC and MR3, this bit can either toggle, go LOW, go HIGH, or do nothing, as selected by MR[11:10]. This bit is driven to the MAT pins if the match function is selected via IOCON. 0 = LOW. 1 = HIGH.

pub fn emc0(&self) -> EMC0R

Bits 4:5 - External Match Control 0. Determines the functionality of External Match 0.

pub fn emc1(&self) -> EMC1R

Bits 6:7 - External Match Control 1. Determines the functionality of External Match 1.

pub fn emc2(&self) -> EMC2R

Bits 8:9 - External Match Control 2. Determines the functionality of External Match 2.

pub fn emc3(&self) -> EMC3R

Bits 10:11 - External Match Control 3. Determines the functionality of External Match 3.