Module lp55s69_pac::dma0::channel::cfg

Configuration register for DMA channel .

Structs

BURSTPOWERR
Value of the field

CHPRIORITYR
Value of the field

R
Value read from the register

W
Value to write to the register

_BURSTPOWERW
Proxy

_CHPRIORITYW
Proxy

_DSTBURSTWRAPW
Proxy

_HWTRIGENW
Proxy

_PERIPHREQENW
Proxy

_SRCBURSTWRAPW
Proxy

_TRIGBURSTW
Proxy

_TRIGPOLW
Proxy

_TRIGTYPEW
Proxy

Enums

DSTBURSTWRAPR
Possible values of the field DSTBURSTWRAP

DSTBURSTWRAPW
Values that can be written to the field DSTBURSTWRAP

HWTRIGENR
Possible values of the field HWTRIGEN

HWTRIGENW
Values that can be written to the field HWTRIGEN

PERIPHREQENR
Possible values of the field PERIPHREQEN

PERIPHREQENW
Values that can be written to the field PERIPHREQEN

SRCBURSTWRAPR
Possible values of the field SRCBURSTWRAP

SRCBURSTWRAPW
Values that can be written to the field SRCBURSTWRAP

TRIGBURSTR
Possible values of the field TRIGBURST

TRIGBURSTW
Values that can be written to the field TRIGBURST

TRIGPOLR
Possible values of the field TRIGPOL

TRIGPOLW
Values that can be written to the field TRIGPOL

TRIGTYPER
Possible values of the field TRIGTYPE

TRIGTYPEW
Values that can be written to the field TRIGTYPE