Module lp55s69_pac::flash_cfpa0

FLASH_CFPA

Modules

cmpa_prog_in_progress
CMPA Page programming on going. This field shall be set to 0x5CC55AA5 in the active CFPA page each time CMPA page programming is going on. It shall always be set to 0x00000000 in the CFPA scratch area.

customer_defined
Customer Defined (Programable through ROM API)

dcfg_cc_socu_dflt
With TZ-M, the part can be sold by level 1 customers (secure code developer) to level-2 customers who develops non-secure code only. - In this scenario, or easy of development, Level-I customer releases the part to always allow non-secure debug. - To allow level-2 customers to further seal the part DCFG_CC_SOCU_NS is used. - ROM will use this word to further restrict the debug access.

dcfg_cc_socu_pin
With TZ-M, the part can be sold by level 1 customers (secure code developer) to level-2 customers who develops non-secure code only. - In this scenario, or easy of development, Level-I customer releases the part to always allow non-secure debug. - To allow level-2 customers to further seal the part DCFG_CC_SOCU_NS is used. - ROM will use this word to further restrict the debug access.

enable_fa_mode
Enable FA mode. SET_FA_MODE Command should write 0xC33CA55A to this word to indicate boot ROM to enter FA mode.

header
.

image_key_revoke
Image key revocation ID (Monotonic counter)

ns_fw_version
Non-Secure firmware version (Monotonic counter)

prince_region0_iv_body0
.

prince_region0_iv_body1
.

prince_region0_iv_body2
.

prince_region0_iv_body3
.

prince_region0_iv_body4
.

prince_region0_iv_body5
.

prince_region0_iv_body6
.

prince_region0_iv_body7
.

prince_region0_iv_body8
.

prince_region0_iv_body9
.

prince_region0_iv_body10
.

prince_region0_iv_body11
.

prince_region0_iv_code0
.

prince_region0_iv_code1
.

prince_region0_iv_code2
.

prince_region0_iv_code3
.

prince_region0_iv_code4
.

prince_region0_iv_code5
.

prince_region0_iv_code6
.

prince_region0_iv_code7
.

prince_region0_iv_code8
.

prince_region0_iv_code9
.

prince_region0_iv_code10
.

prince_region0_iv_code11
.

prince_region0_iv_code12
.

prince_region0_iv_code13
.

prince_region0_iv_header0
.

prince_region0_iv_header1
.

prince_region1_iv_body0
.

prince_region1_iv_body1
.

prince_region1_iv_body2
.

prince_region1_iv_body3
.

prince_region1_iv_body4
.

prince_region1_iv_body5
.

prince_region1_iv_body6
.

prince_region1_iv_body7
.

prince_region1_iv_body8
.

prince_region1_iv_body9
.

prince_region1_iv_body10
.

prince_region1_iv_body11
.

prince_region1_iv_code0
.

prince_region1_iv_code1
.

prince_region1_iv_code2
.

prince_region1_iv_code3
.

prince_region1_iv_code4
.

prince_region1_iv_code5
.

prince_region1_iv_code6
.

prince_region1_iv_code7
.

prince_region1_iv_code8
.

prince_region1_iv_code9
.

prince_region1_iv_code10
.

prince_region1_iv_code11
.

prince_region1_iv_code12
.

prince_region1_iv_code13
.

prince_region1_iv_header0
.

prince_region1_iv_header1
.

prince_region2_iv_body0
.

prince_region2_iv_body1
.

prince_region2_iv_body2
.

prince_region2_iv_body3
.

prince_region2_iv_body4
.

prince_region2_iv_body5
.

prince_region2_iv_body6
.

prince_region2_iv_body7
.

prince_region2_iv_body8
.

prince_region2_iv_body9
.

prince_region2_iv_body10
.

prince_region2_iv_body11
.

prince_region2_iv_code0
.

prince_region2_iv_code1
.

prince_region2_iv_code2
.

prince_region2_iv_code3
.

prince_region2_iv_code4
.

prince_region2_iv_code5
.

prince_region2_iv_code6
.

prince_region2_iv_code7
.

prince_region2_iv_code8
.

prince_region2_iv_code9
.

prince_region2_iv_code10
.

prince_region2_iv_code11
.

prince_region2_iv_code12
.

prince_region2_iv_code13
.

prince_region2_iv_header0
.

prince_region2_iv_header1
.

rotkh_revoke
.

s_fw_version
Secure firmware version (Monotonic counter)

sha256_digest
SHA256_DIGEST0 for DIGEST[31:0] SHA256_DIGEST1 for DIGEST[63:32] SHA256_DIGEST2 for DIGEST[95:64] SHA256_DIGEST3 for DIGEST[127:96] SHA256_DIGEST4 for DIGEST[159:128] SHA256_DIGEST5 for DIGEST[191:160] SHA256_DIGEST6 for DIGEST[223:192] SHA256_DIGEST7 for DIGEST[255:224]

vendor_usage
.

version
.

Structs

CMPA_PROG_IN_PROGRESS
CMPA Page programming on going. This field shall be set to 0x5CC55AA5 in the active CFPA page each time CMPA page programming is going on. It shall always be set to 0x00000000 in the CFPA scratch area.

CUSTOMER_DEFINED
Customer Defined (Programable through ROM API)

DCFG_CC_SOCU_DFLT
With TZ-M, the part can be sold by level 1 customers (secure code developer) to level-2 customers who develops non-secure code only. - In this scenario, or easy of development, Level-I customer releases the part to always allow non-secure debug. - To allow level-2 customers to further seal the part DCFG_CC_SOCU_NS is used. - ROM will use this word to further restrict the debug access.

DCFG_CC_SOCU_PIN
With TZ-M, the part can be sold by level 1 customers (secure code developer) to level-2 customers who develops non-secure code only. - In this scenario, or easy of development, Level-I customer releases the part to always allow non-secure debug. - To allow level-2 customers to further seal the part DCFG_CC_SOCU_NS is used. - ROM will use this word to further restrict the debug access.

ENABLE_FA_MODE
Enable FA mode. SET_FA_MODE Command should write 0xC33CA55A to this word to indicate boot ROM to enter FA mode.

HEADER
.

IMAGE_KEY_REVOKE
Image key revocation ID (Monotonic counter)

NS_FW_VERSION
Non-Secure firmware version (Monotonic counter)

PRINCE_REGION0_IV_BODY0
.

PRINCE_REGION0_IV_BODY1
.

PRINCE_REGION0_IV_BODY2
.

PRINCE_REGION0_IV_BODY3
.

PRINCE_REGION0_IV_BODY4
.

PRINCE_REGION0_IV_BODY5
.

PRINCE_REGION0_IV_BODY6
.

PRINCE_REGION0_IV_BODY7
.

PRINCE_REGION0_IV_BODY8
.

PRINCE_REGION0_IV_BODY9
.

PRINCE_REGION0_IV_BODY10
.

PRINCE_REGION0_IV_BODY11
.

PRINCE_REGION0_IV_CODE0
.

PRINCE_REGION0_IV_CODE1
.

PRINCE_REGION0_IV_CODE2
.

PRINCE_REGION0_IV_CODE3
.

PRINCE_REGION0_IV_CODE4
.

PRINCE_REGION0_IV_CODE5
.

PRINCE_REGION0_IV_CODE6
.

PRINCE_REGION0_IV_CODE7
.

PRINCE_REGION0_IV_CODE8
.

PRINCE_REGION0_IV_CODE9
.

PRINCE_REGION0_IV_CODE10
.

PRINCE_REGION0_IV_CODE11
.

PRINCE_REGION0_IV_CODE12
.

PRINCE_REGION0_IV_CODE13
.

PRINCE_REGION0_IV_HEADER0
.

PRINCE_REGION0_IV_HEADER1
.

PRINCE_REGION1_IV_BODY0
.

PRINCE_REGION1_IV_BODY1
.

PRINCE_REGION1_IV_BODY2
.

PRINCE_REGION1_IV_BODY3
.

PRINCE_REGION1_IV_BODY4
.

PRINCE_REGION1_IV_BODY5
.

PRINCE_REGION1_IV_BODY6
.

PRINCE_REGION1_IV_BODY7
.

PRINCE_REGION1_IV_BODY8
.

PRINCE_REGION1_IV_BODY9
.

PRINCE_REGION1_IV_BODY10
.

PRINCE_REGION1_IV_BODY11
.

PRINCE_REGION1_IV_CODE0
.

PRINCE_REGION1_IV_CODE1
.

PRINCE_REGION1_IV_CODE2
.

PRINCE_REGION1_IV_CODE3
.

PRINCE_REGION1_IV_CODE4
.

PRINCE_REGION1_IV_CODE5
.

PRINCE_REGION1_IV_CODE6
.

PRINCE_REGION1_IV_CODE7
.

PRINCE_REGION1_IV_CODE8
.

PRINCE_REGION1_IV_CODE9
.

PRINCE_REGION1_IV_CODE10
.

PRINCE_REGION1_IV_CODE11
.

PRINCE_REGION1_IV_CODE12
.

PRINCE_REGION1_IV_CODE13
.

PRINCE_REGION1_IV_HEADER0
.

PRINCE_REGION1_IV_HEADER1
.

PRINCE_REGION2_IV_BODY0
.

PRINCE_REGION2_IV_BODY1
.

PRINCE_REGION2_IV_BODY2
.

PRINCE_REGION2_IV_BODY3
.

PRINCE_REGION2_IV_BODY4
.

PRINCE_REGION2_IV_BODY5
.

PRINCE_REGION2_IV_BODY6
.

PRINCE_REGION2_IV_BODY7
.

PRINCE_REGION2_IV_BODY8
.

PRINCE_REGION2_IV_BODY9
.

PRINCE_REGION2_IV_BODY10
.

PRINCE_REGION2_IV_BODY11
.

PRINCE_REGION2_IV_CODE0
.

PRINCE_REGION2_IV_CODE1
.

PRINCE_REGION2_IV_CODE2
.

PRINCE_REGION2_IV_CODE3
.

PRINCE_REGION2_IV_CODE4
.

PRINCE_REGION2_IV_CODE5
.

PRINCE_REGION2_IV_CODE6
.

PRINCE_REGION2_IV_CODE7
.

PRINCE_REGION2_IV_CODE8
.

PRINCE_REGION2_IV_CODE9
.

PRINCE_REGION2_IV_CODE10
.

PRINCE_REGION2_IV_CODE11
.

PRINCE_REGION2_IV_CODE12
.

PRINCE_REGION2_IV_CODE13
.

PRINCE_REGION2_IV_HEADER0
.

PRINCE_REGION2_IV_HEADER1
.

ROTKH_REVOKE
.

RegisterBlock
Register block

SHA256_DIGEST
SHA256_DIGEST0 for DIGEST[31:0] SHA256_DIGEST1 for DIGEST[63:32] SHA256_DIGEST2 for DIGEST[95:64] SHA256_DIGEST3 for DIGEST[127:96] SHA256_DIGEST4 for DIGEST[159:128] SHA256_DIGEST5 for DIGEST[191:160] SHA256_DIGEST6 for DIGEST[223:192] SHA256_DIGEST7 for DIGEST[255:224]

S_FW_VERSION
Secure firmware version (Monotonic counter)

VENDOR_USAGE
.

VERSION
.