Module lp55s69_pac::hashcrypt::cryptcfg

Crypto settings for AES and Salsa and ChaCha

Structs

AESCTRPOSR
Value of the field

MSW1ST_OUTR
Value of the field

MSW1STR
Value of the field

R
Value read from the register

STREAMLASTR
Value of the field

SWAPDATR
Value of the field

SWAPKEYR
Value of the field

W
Value to write to the register

XSALSAR
Value of the field

_AESCTRPOSW
Proxy

_AESDECRYPTW
Proxy

_AESKEYSZW
Proxy

_AESMODEW
Proxy

_AESSECRETW
Proxy

_ICBSTRMW
Proxy

_ICBSZW
Proxy

_MSW1ST_OUTW
Proxy

_MSW1STW
Proxy

_STREAMLASTW
Proxy

_SWAPDATW
Proxy

_SWAPKEYW
Proxy

_XSALSAW
Proxy

Enums

AESDECRYPTR
Possible values of the field AESDECRYPT

AESDECRYPTW
Values that can be written to the field AESDECRYPT

AESKEYSZR
Possible values of the field AESKEYSZ

AESKEYSZW
Values that can be written to the field AESKEYSZ

AESMODER
Possible values of the field AESMODE

AESMODEW
Values that can be written to the field AESMODE

AESSECRETR
Possible values of the field AESSECRET

AESSECRETW
Values that can be written to the field AESSECRET

ICBSTRMR
Possible values of the field ICBSTRM

ICBSTRMW
Values that can be written to the field ICBSTRM

ICBSZR
Possible values of the field ICBSZ

ICBSZW
Values that can be written to the field ICBSZ