Module lp55s69_pac::hashcrypt::lock

Lock register allows locking to the current security level or unlocking by the lock holding level.

Structs

PATTERNR
Value of the field

R
Value read from the register

W
Value to write to the register

_PATTERNW
Proxy

_SECLOCKW
Proxy

Enums

SECLOCKR
Possible values of the field SECLOCK

SECLOCKW
Values that can be written to the field SECLOCK