Module lp55s69_pac::i2s0::fifocfg

FIFO configuration and enable register.

Structs

EMPTYRXR
Value of the field

EMPTYTXR
Value of the field

R
Value read from the register

SIZER
Value of the field

W
Value to write to the register

_DMARXW
Proxy

_DMATXW
Proxy

_EMPTYRXW
Proxy

_EMPTYTXW
Proxy

_ENABLERXW
Proxy

_ENABLETXW
Proxy

_PACK48W
Proxy

_POPDBGW
Proxy

_TXI2SE0W
Proxy

_WAKERXW
Proxy

_WAKETXW
Proxy

Enums

DMARXR
Possible values of the field DMARX

DMARXW
Values that can be written to the field DMARX

DMATXR
Possible values of the field DMATX

DMATXW
Values that can be written to the field DMATX

ENABLERXR
Possible values of the field ENABLERX

ENABLERXW
Values that can be written to the field ENABLERX

ENABLETXR
Possible values of the field ENABLETX

ENABLETXW
Values that can be written to the field ENABLETX

PACK48R
Possible values of the field PACK48

PACK48W
Values that can be written to the field PACK48

POPDBGR
Possible values of the field POPDBG

POPDBGW
Values that can be written to the field POPDBG

TXI2SE0R
Possible values of the field TXI2SE0

TXI2SE0W
Values that can be written to the field TXI2SE0

WAKERXR
Possible values of the field WAKERX

WAKERXW
Values that can be written to the field WAKERX

WAKETXR
Possible values of the field WAKETX

WAKETXW
Values that can be written to the field WAKETX