Enum lp55s69_pac::inputmux::dma1_itrig_inmux::INPR

pub enum INPR {
    VAL0,
    VAL1,
    VAL2,
    VAL3,
    VAL4,
    VAL5,
    VAL6,
    VAL7,
    VAL8,
    VAL9,
    VAL10,
    VAL11,
    VAL12,
    VAL13,
    VAL14,
    VAL15,
}

Possible values of the field INP

Variants
VAL0

Pin interrupt 0

VAL1

Pin interrupt 1

VAL2

Pin interrupt 2

VAL3

Pin interrupt 3

VAL4

Timer CTIMER0 Match 0

VAL5

Timer CTIMER0 Match 1

VAL6

Timer CTIMER2 Match 0

VAL7

Timer CTIMER4 Match 0

VAL8

DMA1 output trigger mux 0

VAL9

DMA1 output trigger mux 1

VAL10

DMA1 output trigger mux 2

VAL11

DMA1 output trigger mux 3

VAL12

SCT0 DMA request 0

VAL13

SCT0 DMA request 1

VAL14

HASH DMA RX trigger

VAL15

None

Methods

impl INPR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_val0(&self) -> bool

Checks if the value of the field is VAL0

pub fn is_val1(&self) -> bool

Checks if the value of the field is VAL1

pub fn is_val2(&self) -> bool

Checks if the value of the field is VAL2

pub fn is_val3(&self) -> bool

Checks if the value of the field is VAL3

pub fn is_val4(&self) -> bool

Checks if the value of the field is VAL4

pub fn is_val5(&self) -> bool

Checks if the value of the field is VAL5

pub fn is_val6(&self) -> bool

Checks if the value of the field is VAL6

pub fn is_val7(&self) -> bool

Checks if the value of the field is VAL7

pub fn is_val8(&self) -> bool

Checks if the value of the field is VAL8

pub fn is_val9(&self) -> bool

Checks if the value of the field is VAL9

pub fn is_val10(&self) -> bool

Checks if the value of the field is VAL10

pub fn is_val11(&self) -> bool

Checks if the value of the field is VAL11

pub fn is_val12(&self) -> bool

Checks if the value of the field is VAL12

pub fn is_val13(&self) -> bool

Checks if the value of the field is VAL13

pub fn is_val14(&self) -> bool

Checks if the value of the field is VAL14

pub fn is_val15(&self) -> bool

Checks if the value of the field is VAL15