Module lp55s69_pac::iocon::pio1_10

Digital I/O control for port 1 pins PIO1_10

Structs

R
Value read from the register

W
Value to write to the register

_DIGIMODEW
Proxy

_FUNCW
Proxy

_INVERTW
Proxy

_MODEW
Proxy

_ODW
Proxy

_SLEWW
Proxy

Enums

DIGIMODER
Possible values of the field DIGIMODE

DIGIMODEW
Values that can be written to the field DIGIMODE

FUNCR
Possible values of the field FUNC

FUNCW
Values that can be written to the field FUNC

INVERTR
Possible values of the field INVERT

INVERTW
Values that can be written to the field INVERT

MODER
Possible values of the field MODE

MODEW
Values that can be written to the field MODE

ODR
Possible values of the field OD

ODW
Values that can be written to the field OD

SLEWR
Possible values of the field SLEW

SLEWW
Values that can be written to the field SLEW