Module lp55s69_pac::pint::pmctrl

Pattern match interrupt control register

Structs

PMATR
Value of the field

R
Value read from the register

W
Value to write to the register

_ENA_RXEVW
Proxy

_PMATW
Proxy

_SEL_PMATCHW
Proxy

Enums

ENA_RXEVR
Possible values of the field ENA_RXEV

ENA_RXEVW
Values that can be written to the field ENA_RXEV

SEL_PMATCHR
Possible values of the field SEL_PMATCH

SEL_PMATCHW
Values that can be written to the field SEL_PMATCH