Module lp55s69_pac::pint::pmsrc

Pattern match interrupt bit-slice source register

Structs

R
Value read from the register

W
Value to write to the register

_SRC0W
Proxy

_SRC1W
Proxy

_SRC2W
Proxy

_SRC3W
Proxy

_SRC4W
Proxy

_SRC5W
Proxy

_SRC6W
Proxy

_SRC7W
Proxy

Enums

SRC0R
Possible values of the field SRC0

SRC0W
Values that can be written to the field SRC0

SRC1R
Possible values of the field SRC1

SRC1W
Values that can be written to the field SRC1

SRC2R
Possible values of the field SRC2

SRC2W
Values that can be written to the field SRC2

SRC3R
Possible values of the field SRC3

SRC3W
Values that can be written to the field SRC3

SRC4R
Possible values of the field SRC4

SRC4W
Values that can be written to the field SRC4

SRC5R
Possible values of the field SRC5

SRC5W
Values that can be written to the field SRC5

SRC6R
Possible values of the field SRC6

SRC6W
Values that can be written to the field SRC6

SRC7R
Possible values of the field SRC7

SRC7W
Values that can be written to the field SRC7