Enum lp55s69_pac::pmc::bodvbat::TRIGLVLR

pub enum TRIGLVLR {
    V_1P00,
    V_1P10,
    V_1P20,
    V_1P30,
    V_1P40,
    V_1P50,
    V_1P60,
    V_1P65,
    V_1P70,
    V_1P75,
    V_1P80,
    V_1P90,
    V_2P00,
    V_2P10,
    V_2P20,
    V_2P30,
    V_2P40,
    V_2P50,
    V_2P60,
    V_2P70,
    V_2P80,
    V_2P90,
    V_3P00,
    V_3P10,
    V_3P20,
    V_3P30_2,
    V_3P30_3,
    V_3P30_4,
    V_3P30_5,
    V_3P30_6,
    V_3P30_7,
    V_3P30_8,
}

Possible values of the field TRIGLVL

Variants
V_1P00

1.00 V.

V_1P10

1.10 V.

V_1P20

1.20 V.

V_1P30

1.30 V.

V_1P40

1.40 V.

V_1P50

1.50 V.

V_1P60

1.60 V.

V_1P65

1.65 V.

V_1P70

1.70 V.

V_1P75

1.75 V.

V_1P80

1.80 V.

V_1P90

1.90 V.

V_2P00

2.00 V.

V_2P10

2.10 V.

V_2P20

2.20 V.

V_2P30

2.30 V.

V_2P40

2.40 V.

V_2P50

2.50 V.

V_2P60

2.60 V.

V_2P70

2.70 V.

V_2P80

2.806 V.

V_2P90

2.90 V.

V_3P00

3.00 V.

V_3P10

3.10 V.

V_3P20

3.20 V.

V_3P30_2

3.30 V.

V_3P30_3

3.30 V.

V_3P30_4

3.30 V.

V_3P30_5

3.30 V.

V_3P30_6

3.30 V.

V_3P30_7

3.30 V.

V_3P30_8

3.30 V.

Methods

impl TRIGLVLR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_v_1p00(&self) -> bool

Checks if the value of the field is V_1P00

pub fn is_v_1p10(&self) -> bool

Checks if the value of the field is V_1P10

pub fn is_v_1p20(&self) -> bool

Checks if the value of the field is V_1P20

pub fn is_v_1p30(&self) -> bool

Checks if the value of the field is V_1P30

pub fn is_v_1p40(&self) -> bool

Checks if the value of the field is V_1P40

pub fn is_v_1p50(&self) -> bool

Checks if the value of the field is V_1P50

pub fn is_v_1p60(&self) -> bool

Checks if the value of the field is V_1P60

pub fn is_v_1p65(&self) -> bool

Checks if the value of the field is V_1P65

pub fn is_v_1p70(&self) -> bool

Checks if the value of the field is V_1P70

pub fn is_v_1p75(&self) -> bool

Checks if the value of the field is V_1P75

pub fn is_v_1p80(&self) -> bool

Checks if the value of the field is V_1P80

pub fn is_v_1p90(&self) -> bool

Checks if the value of the field is V_1P90

pub fn is_v_2p00(&self) -> bool

Checks if the value of the field is V_2P00

pub fn is_v_2p10(&self) -> bool

Checks if the value of the field is V_2P10

pub fn is_v_2p20(&self) -> bool

Checks if the value of the field is V_2P20

pub fn is_v_2p30(&self) -> bool

Checks if the value of the field is V_2P30

pub fn is_v_2p40(&self) -> bool

Checks if the value of the field is V_2P40

pub fn is_v_2p50(&self) -> bool

Checks if the value of the field is V_2P50

pub fn is_v_2p60(&self) -> bool

Checks if the value of the field is V_2P60

pub fn is_v_2p70(&self) -> bool

Checks if the value of the field is V_2P70

pub fn is_v_2p80(&self) -> bool

Checks if the value of the field is V_2P80

pub fn is_v_2p90(&self) -> bool

Checks if the value of the field is V_2P90

pub fn is_v_3p00(&self) -> bool

Checks if the value of the field is V_3P00

pub fn is_v_3p10(&self) -> bool

Checks if the value of the field is V_3P10

pub fn is_v_3p20(&self) -> bool

Checks if the value of the field is V_3P20

pub fn is_v_3p30_2(&self) -> bool

Checks if the value of the field is V_3P30_2

pub fn is_v_3p30_3(&self) -> bool

Checks if the value of the field is V_3P30_3

pub fn is_v_3p30_4(&self) -> bool

Checks if the value of the field is V_3P30_4

pub fn is_v_3p30_5(&self) -> bool

Checks if the value of the field is V_3P30_5

pub fn is_v_3p30_6(&self) -> bool

Checks if the value of the field is V_3P30_6

pub fn is_v_3p30_7(&self) -> bool

Checks if the value of the field is V_3P30_7

pub fn is_v_3p30_8(&self) -> bool

Checks if the value of the field is V_3P30_8