Struct lp55s69_pac::pmc::resetcause::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn por(&mut self) -> _PORW

Bit 0 - 1 : The last chip reset was caused by a Power On Reset. Write '1' to clear this bit.

pub fn padreset(&mut self) -> _PADRESETW

Bit 1 - 1 : The last chip reset was caused by a Pin Reset. Write '1' to clear this bit.

pub fn bodreset(&mut self) -> _BODRESETW

Bit 2 - 1 : The last chip reset was caused by a Brown Out Detector (BoD), either VBAT BoD or Core Logic BoD. Write '1' to clear this bit.

pub fn systemreset(&mut self) -> _SYSTEMRESETW

Bit 3 - 1 : The last chip reset was caused by a System Reset requested by the ARM CPU. Write '1' to clear this bit.

pub fn wdtreset(&mut self) -> _WDTRESETW

Bit 4 - 1 : The last chip reset was caused by the Watchdog Timer. Write '1' to clear this bit.

pub fn swrreset(&mut self) -> _SWRRESETW

Bit 5 - 1 : The last chip reset was caused by a Software. Write '1' to clear this bit.

pub fn dpdreset_wakeupio(&mut self) -> _DPDRESET_WAKEUPIOW

Bit 6 - 1 : The last chip reset was caused by a Wake-up I/O reset event during DEEP POWER DOWN mode. Write '1' to clear this bit.

pub fn dpdreset_rtc(&mut self) -> _DPDRESET_RTCW

Bit 7 - 1 : The last chip reset was caused by an RTC (either RTC Alarm or RTC wake up) reset event during DEEP POWER DOWN mode. Write '1' to clear this bit.

pub fn dpdreset_ostimer(&mut self) -> _DPDRESET_OSTIMERW

Bit 8 - 1 : The last chip reset was caused by a OS Event Timer reset eventduring DEEP POWER DOWN mode. Write '1' to clear this bit.