Module lp55s69_pac::rtc::ctrl

RTC control register

Structs

R
Value read from the register

W
Value to write to the register

_ALARM1HZW
Proxy

_ALARMDPD_ENW
Proxy

_RTC1KHZ_ENW
Proxy

_RTC_ENW
Proxy

_RTC_OSC_BYPASSW
Proxy

_RTC_OSC_PDW
Proxy

_RTC_SUBSEC_ENAW
Proxy

_SWRESETW
Proxy

_WAKE1KHZW
Proxy

_WAKEDPD_ENW
Proxy

Enums

ALARM1HZR
Possible values of the field ALARM1HZ

ALARM1HZW
Values that can be written to the field ALARM1HZ

ALARMDPD_ENR
Possible values of the field ALARMDPD_EN

ALARMDPD_ENW
Values that can be written to the field ALARMDPD_EN

RTC1KHZ_ENR
Possible values of the field RTC1KHZ_EN

RTC1KHZ_ENW
Values that can be written to the field RTC1KHZ_EN

RTC_ENR
Possible values of the field RTC_EN

RTC_ENW
Values that can be written to the field RTC_EN

RTC_OSC_BYPASSR
Possible values of the field RTC_OSC_BYPASS

RTC_OSC_BYPASSW
Values that can be written to the field RTC_OSC_BYPASS

RTC_OSC_PDR
Possible values of the field RTC_OSC_PD

RTC_OSC_PDW
Values that can be written to the field RTC_OSC_PD

RTC_SUBSEC_ENAR
Possible values of the field RTC_SUBSEC_ENA

RTC_SUBSEC_ENAW
Values that can be written to the field RTC_SUBSEC_ENA

SWRESETR
Possible values of the field SWRESET

SWRESETW
Values that can be written to the field SWRESET

WAKE1KHZR
Possible values of the field WAKE1KHZ

WAKE1KHZW
Values that can be written to the field WAKE1KHZ

WAKEDPD_ENR
Possible values of the field WAKEDPD_EN

WAKEDPD_ENW
Values that can be written to the field WAKEDPD_EN