Module lp55s69_pac::scn_scb::cppwr

Coprocessor Power Control Register

Structs

R
Value read from the register

SU11R
Value of the field

SUS11R
Value of the field

W
Value to write to the register

_SU0W
Proxy

_SU1W
Proxy

_SU2W
Proxy

_SU3W
Proxy

_SU4W
Proxy

_SU5W
Proxy

_SU6W
Proxy

_SU7W
Proxy

_SU10W
Proxy

_SU11W
Proxy

_SUS0W
Proxy

_SUS1W
Proxy

_SUS2W
Proxy

_SUS3W
Proxy

_SUS4W
Proxy

_SUS5W
Proxy

_SUS6W
Proxy

_SUS7W
Proxy

_SUS10W
Proxy

_SUS11W
Proxy

Enums

SU0R
Possible values of the field SU0

SU0W
Values that can be written to the field SU0

SU1R
Possible values of the field SU1

SU1W
Values that can be written to the field SU1

SU2R
Possible values of the field SU2

SU2W
Values that can be written to the field SU2

SU3R
Possible values of the field SU3

SU3W
Values that can be written to the field SU3

SU4R
Possible values of the field SU4

SU4W
Values that can be written to the field SU4

SU5R
Possible values of the field SU5

SU5W
Values that can be written to the field SU5

SU6R
Possible values of the field SU6

SU6W
Values that can be written to the field SU6

SU7R
Possible values of the field SU7

SU7W
Values that can be written to the field SU7

SU10R
Possible values of the field SU10

SU10W
Values that can be written to the field SU10

SUS0R
Possible values of the field SUS0

SUS0W
Values that can be written to the field SUS0

SUS1R
Possible values of the field SUS1

SUS1W
Values that can be written to the field SUS1

SUS2R
Possible values of the field SUS2

SUS2W
Values that can be written to the field SUS2

SUS3R
Possible values of the field SUS3

SUS3W
Values that can be written to the field SUS3

SUS4R
Possible values of the field SUS4

SUS4W
Values that can be written to the field SUS4

SUS5R
Possible values of the field SUS5

SUS5W
Values that can be written to the field SUS5

SUS6R
Possible values of the field SUS6

SUS6W
Values that can be written to the field SUS6

SUS7R
Possible values of the field SUS7

SUS7W
Values that can be written to the field SUS7

SUS10R
Possible values of the field SUS10

SUS10W
Values that can be written to the field SUS10