Module lp55s69_pac::sct0::outputdirctrl

SCT output counter direction control register

Structs

R
Value read from the register

W
Value to write to the register

_SETCLR0W
Proxy

_SETCLR1W
Proxy

_SETCLR2W
Proxy

_SETCLR3W
Proxy

_SETCLR4W
Proxy

_SETCLR5W
Proxy

_SETCLR6W
Proxy

_SETCLR7W
Proxy

_SETCLR8W
Proxy

_SETCLR9W
Proxy

_SETCLR10W
Proxy

_SETCLR11W
Proxy

_SETCLR12W
Proxy

_SETCLR13W
Proxy

_SETCLR14W
Proxy

_SETCLR15W
Proxy

Enums

SETCLR0R
Possible values of the field SETCLR0

SETCLR0W
Values that can be written to the field SETCLR0

SETCLR1R
Possible values of the field SETCLR1

SETCLR1W
Values that can be written to the field SETCLR1

SETCLR2R
Possible values of the field SETCLR2

SETCLR2W
Values that can be written to the field SETCLR2

SETCLR3R
Possible values of the field SETCLR3

SETCLR3W
Values that can be written to the field SETCLR3

SETCLR4R
Possible values of the field SETCLR4

SETCLR4W
Values that can be written to the field SETCLR4

SETCLR5R
Possible values of the field SETCLR5

SETCLR5W
Values that can be written to the field SETCLR5

SETCLR6R
Possible values of the field SETCLR6

SETCLR6W
Values that can be written to the field SETCLR6

SETCLR7R
Possible values of the field SETCLR7

SETCLR7W
Values that can be written to the field SETCLR7

SETCLR8R
Possible values of the field SETCLR8

SETCLR8W
Values that can be written to the field SETCLR8

SETCLR9R
Possible values of the field SETCLR9

SETCLR9W
Values that can be written to the field SETCLR9

SETCLR10R
Possible values of the field SETCLR10

SETCLR10W
Values that can be written to the field SETCLR10

SETCLR11R
Possible values of the field SETCLR11

SETCLR11W
Values that can be written to the field SETCLR11

SETCLR12R
Possible values of the field SETCLR12

SETCLR12W
Values that can be written to the field SETCLR12

SETCLR13R
Possible values of the field SETCLR13

SETCLR13W
Values that can be written to the field SETCLR13

SETCLR14R
Possible values of the field SETCLR14

SETCLR14W
Values that can be written to the field SETCLR14

SETCLR15R
Possible values of the field SETCLR15

SETCLR15W
Values that can be written to the field SETCLR15