Struct lp55s69_pac::sdif::cmd::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn cmd_index(&self) -> CMD_INDEXR

Bits 0:5 - Command index.

pub fn response_expect(&self) -> RESPONSE_EXPECTR

Bit 6 - Response expect.

pub fn response_length(&self) -> RESPONSE_LENGTHR

Bit 7 - Response length.

pub fn check_response_crc(&self) -> CHECK_RESPONSE_CRCR

Bit 8 - Check response CRC.

pub fn data_expected(&self) -> DATA_EXPECTEDR

Bit 9 - Data expected.

pub fn read_write(&self) -> READ_WRITER

Bit 10 - read/write.

pub fn transfer_mode(&self) -> TRANSFER_MODER

Bit 11 - Transfer mode.

pub fn send_auto_stop(&self) -> SEND_AUTO_STOPR

Bit 12 - Send auto stop.

pub fn wait_prvdata_complete(&self) -> WAIT_PRVDATA_COMPLETER

Bit 13 - Wait prvdata complete.

pub fn stop_abort_cmd(&self) -> STOP_ABORT_CMDR

Bit 14 - Stop abort command.

pub fn send_initialization(&self) -> SEND_INITIALIZATIONR

Bit 15 - Send initialization.

pub fn card_number(&self) -> CARD_NUMBERR

Bits 16:20 - Specifies the card number of SDCARD for which the current Command is being executed

pub fn update_clock_registers_only(&self) -> UPDATE_CLOCK_REGISTERS_ONLYR

Bit 21 - Update clock registers only.

pub fn read_ceata_device(&self) -> READ_CEATA_DEVICER

Bit 22 - Read ceata device.

pub fn ccs_expected(&self) -> CCS_EXPECTEDR

Bit 23 - CCS expected.

pub fn enable_boot(&self) -> ENABLE_BOOTR

Bit 24 - Enable Boot - this bit should be set only for mandatory boot mode.

pub fn expect_boot_ack(&self) -> EXPECT_BOOT_ACKR

Bit 25 - Expect Boot Acknowledge.

pub fn disable_boot(&self) -> DISABLE_BOOTR

Bit 26 - Disable Boot.

pub fn boot_mode(&self) -> BOOT_MODER

Bit 27 - Boot Mode.

pub fn volt_switch(&self) -> VOLT_SWITCHR

Bit 28 - Voltage switch bit.

pub fn use_hold_reg(&self) -> USE_HOLD_REGR

Bit 29 - Use Hold Register.

pub fn start_cmd(&self) -> START_CMDR

Bit 31 - Start command.