Struct lp55s69_pac::sdif::cmd::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn cmd_index(&mut self) -> _CMD_INDEXW

Bits 0:5 - Command index.

pub fn response_expect(&mut self) -> _RESPONSE_EXPECTW

Bit 6 - Response expect.

pub fn response_length(&mut self) -> _RESPONSE_LENGTHW

Bit 7 - Response length.

pub fn check_response_crc(&mut self) -> _CHECK_RESPONSE_CRCW

Bit 8 - Check response CRC.

pub fn data_expected(&mut self) -> _DATA_EXPECTEDW

Bit 9 - Data expected.

pub fn read_write(&mut self) -> _READ_WRITEW

Bit 10 - read/write.

pub fn transfer_mode(&mut self) -> _TRANSFER_MODEW

Bit 11 - Transfer mode.

pub fn send_auto_stop(&mut self) -> _SEND_AUTO_STOPW

Bit 12 - Send auto stop.

pub fn wait_prvdata_complete(&mut self) -> _WAIT_PRVDATA_COMPLETEW

Bit 13 - Wait prvdata complete.

pub fn stop_abort_cmd(&mut self) -> _STOP_ABORT_CMDW

Bit 14 - Stop abort command.

pub fn send_initialization(&mut self) -> _SEND_INITIALIZATIONW

Bit 15 - Send initialization.

pub fn card_number(&mut self) -> _CARD_NUMBERW

Bits 16:20 - Specifies the card number of SDCARD for which the current Command is being executed

pub fn update_clock_registers_only(&mut self) -> _UPDATE_CLOCK_REGISTERS_ONLYW

Bit 21 - Update clock registers only.

pub fn read_ceata_device(&mut self) -> _READ_CEATA_DEVICEW

Bit 22 - Read ceata device.

pub fn ccs_expected(&mut self) -> _CCS_EXPECTEDW

Bit 23 - CCS expected.

pub fn enable_boot(&mut self) -> _ENABLE_BOOTW

Bit 24 - Enable Boot - this bit should be set only for mandatory boot mode.

pub fn expect_boot_ack(&mut self) -> _EXPECT_BOOT_ACKW

Bit 25 - Expect Boot Acknowledge.

pub fn disable_boot(&mut self) -> _DISABLE_BOOTW

Bit 26 - Disable Boot.

pub fn boot_mode(&mut self) -> _BOOT_MODEW

Bit 27 - Boot Mode.

pub fn volt_switch(&mut self) -> _VOLT_SWITCHW

Bit 28 - Voltage switch bit.

pub fn use_hold_reg(&mut self) -> _USE_HOLD_REGW

Bit 29 - Use Hold Register.

pub fn start_cmd(&mut self) -> _START_CMDW

Bit 31 - Start command.