Module lp55s69_pac::spi0::cfg

SPI Configuration register

Structs

R
Value read from the register

W
Value to write to the register

_CPHAW
Proxy

_CPOLW
Proxy

_ENABLEW
Proxy

_LOOPW
Proxy

_LSBFW
Proxy

_MASTERW
Proxy

_SPOL0W
Proxy

_SPOL1W
Proxy

_SPOL2W
Proxy

_SPOL3W
Proxy

Enums

CPHAR
Possible values of the field CPHA

CPHAW
Values that can be written to the field CPHA

CPOLR
Possible values of the field CPOL

CPOLW
Values that can be written to the field CPOL

ENABLER
Possible values of the field ENABLE

ENABLEW
Values that can be written to the field ENABLE

LOOPR
Possible values of the field LOOP

LOOPW
Values that can be written to the field LOOP

LSBFR
Possible values of the field LSBF

LSBFW
Values that can be written to the field LSBF

MASTERR
Possible values of the field MASTER

MASTERW
Values that can be written to the field MASTER

SPOL0R
Possible values of the field SPOL0

SPOL0W
Values that can be written to the field SPOL0

SPOL1R
Possible values of the field SPOL1

SPOL1W
Values that can be written to the field SPOL1

SPOL2R
Possible values of the field SPOL2

SPOL2W
Values that can be written to the field SPOL2

SPOL3R
Possible values of the field SPOL3

SPOL3W
Values that can be written to the field SPOL3