Enum lp55s69_pac::syscon::adcclksel::SELR

pub enum SELR {
    ENUM_0X0,
    ENUM_0X1,
    ENUM_0X2,
    ENUM_0X3,
    ENUM_0X4,
    ENUM_0X5,
    ENUM_0X6,
    ENUM_0X7,
}

Possible values of the field SEL

Variants
ENUM_0X0

Main clock.

ENUM_0X1

PLL0 clock.

ENUM_0X2

FRO 96 MHz clock.

ENUM_0X3

No clock.

ENUM_0X4

No clock.

ENUM_0X5

No clock.

ENUM_0X6

No clock.

ENUM_0X7

No clock.

Methods

impl SELR

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_enum_0x0(&self) -> bool

Checks if the value of the field is ENUM_0X0

pub fn is_enum_0x1(&self) -> bool

Checks if the value of the field is ENUM_0X1

pub fn is_enum_0x2(&self) -> bool

Checks if the value of the field is ENUM_0X2

pub fn is_enum_0x3(&self) -> bool

Checks if the value of the field is ENUM_0X3

pub fn is_enum_0x4(&self) -> bool

Checks if the value of the field is ENUM_0X4

pub fn is_enum_0x5(&self) -> bool

Checks if the value of the field is ENUM_0X5

pub fn is_enum_0x6(&self) -> bool

Checks if the value of the field is ENUM_0X6

pub fn is_enum_0x7(&self) -> bool

Checks if the value of the field is ENUM_0X7