Module lp55s69_pac::syscon::ahbclkctrl1

AHB Clock control 1

Structs

R
Value read from the register

W
Value to write to the register

_EZHAW
Proxy

_EZHBW
Proxy

_FC0W
Proxy

_FC1W
Proxy

_FC2W
Proxy

_FC3W
Proxy

_FC4W
Proxy

_FC5W
Proxy

_FC6W
Proxy

_FC7W
Proxy

_MRTW
Proxy

_OSTIMER0W
Proxy

_PVTW
Proxy

_SCT0W
Proxy

_SCTIPUW
Proxy

_TIMER0W
Proxy

_TIMER1W
Proxy

_TIMER2W
Proxy

_USB0_DEVW
Proxy

_UTICK0W
Proxy

Enums

EZHAR
Possible values of the field EZHA

EZHAW
Values that can be written to the field EZHA

EZHBR
Possible values of the field EZHB

EZHBW
Values that can be written to the field EZHB

FC0R
Possible values of the field FC0

FC0W
Values that can be written to the field FC0

FC1R
Possible values of the field FC1

FC1W
Values that can be written to the field FC1

FC2R
Possible values of the field FC2

FC2W
Values that can be written to the field FC2

FC3R
Possible values of the field FC3

FC3W
Values that can be written to the field FC3

FC4R
Possible values of the field FC4

FC4W
Values that can be written to the field FC4

FC5R
Possible values of the field FC5

FC5W
Values that can be written to the field FC5

FC6R
Possible values of the field FC6

FC6W
Values that can be written to the field FC6

FC7R
Possible values of the field FC7

FC7W
Values that can be written to the field FC7

MRTR
Possible values of the field MRT

MRTW
Values that can be written to the field MRT

OSTIMER0R
Possible values of the field OSTIMER0

OSTIMER0W
Values that can be written to the field OSTIMER0

PVTR
Possible values of the field PVT

PVTW
Values that can be written to the field PVT

SCT0R
Possible values of the field SCT0

SCT0W
Values that can be written to the field SCT0

SCTIPUR
Possible values of the field SCTIPU

SCTIPUW
Values that can be written to the field SCTIPU

TIMER0R
Possible values of the field TIMER0

TIMER0W
Values that can be written to the field TIMER0

TIMER1R
Possible values of the field TIMER1

TIMER1W
Values that can be written to the field TIMER1

TIMER2R
Possible values of the field TIMER2

TIMER2W
Values that can be written to the field TIMER2

USB0_DEVR
Possible values of the field USB0_DEV

USB0_DEVW
Values that can be written to the field USB0_DEV

UTICK0R
Possible values of the field UTICK0

UTICK0W
Values that can be written to the field UTICK0