Module lp55s69_pac::syscon::autoclkgateoverride

Control automatic clock gating

Structs

R
Value read from the register

W
Value to write to the register

_CRCGENW
Proxy

_ENABLEUPDATEW
Proxy

_FLASHW
Proxy

_FMCW
Proxy

_RAM0_CTRLW
Proxy

_RAM1_CTRLW
Proxy

_RAM2_CTRLW
Proxy

_RAM3_CTRLW
Proxy

_RAM4_CTRLW
Proxy

_RAMX_CTRLW
Proxy

_ROMW
Proxy

_SDMA0W
Proxy

_SDMA1W
Proxy

_SYNC0_APBW
Proxy

_SYNC1_APBW
Proxy

_SYSCONW
Proxy

_USBW
Proxy

Enums

CRCGENR
Possible values of the field CRCGEN

CRCGENW
Values that can be written to the field CRCGEN

ENABLEUPDATEW
Values that can be written to the field ENABLEUPDATE

FLASHR
Possible values of the field FLASH

FLASHW
Values that can be written to the field FLASH

FMCR
Possible values of the field FMC

FMCW
Values that can be written to the field FMC

RAM0_CTRLR
Possible values of the field RAM0_CTRL

RAM0_CTRLW
Values that can be written to the field RAM0_CTRL

RAM1_CTRLR
Possible values of the field RAM1_CTRL

RAM1_CTRLW
Values that can be written to the field RAM1_CTRL

RAM2_CTRLR
Possible values of the field RAM2_CTRL

RAM2_CTRLW
Values that can be written to the field RAM2_CTRL

RAM3_CTRLR
Possible values of the field RAM3_CTRL

RAM3_CTRLW
Values that can be written to the field RAM3_CTRL

RAM4_CTRLR
Possible values of the field RAM4_CTRL

RAM4_CTRLW
Values that can be written to the field RAM4_CTRL

RAMX_CTRLR
Possible values of the field RAMX_CTRL

RAMX_CTRLW
Values that can be written to the field RAMX_CTRL

ROMR
Possible values of the field ROM

ROMW
Values that can be written to the field ROM

SDMA0R
Possible values of the field SDMA0

SDMA0W
Values that can be written to the field SDMA0

SDMA1R
Possible values of the field SDMA1

SDMA1W
Values that can be written to the field SDMA1

SYNC0_APBR
Possible values of the field SYNC0_APB

SYNC0_APBW
Values that can be written to the field SYNC0_APB

SYNC1_APBR
Possible values of the field SYNC1_APB

SYNC1_APBW
Values that can be written to the field SYNC1_APB

SYSCONR
Possible values of the field SYSCON

SYSCONW
Values that can be written to the field SYSCON

USBR
Possible values of the field USB

USBW
Values that can be written to the field USB