Module lp55s69_pac::syscon::clock_ctrl

Various system clock controls : Flash clock (48 MHz) control, clocks to Frequency Measures

Structs

R
Value read from the register

W
Value to write to the register

_ANA_FRO12M_CLK_ENAW
Proxy

_CLKIN_ENAW
Proxy

_FLASH48MHZ_ENAW
Proxy

_FRO1MHZ_UTICK_ENAW
Proxy

_FRO1MHZ_CLK_ENAW
Proxy

_FRO12MHZ_FREQM_ENAW
Proxy

_FRO_HF_FREQM_ENAW
Proxy

_PLU_DEGLITCH_CLK_ENAW
Proxy

_XO_CAL_CLK_ENAW
Proxy

_XTAL32MHZ_FREQM_ENAW
Proxy

Enums

ANA_FRO12M_CLK_ENAR
Possible values of the field ANA_FRO12M_CLK_ENA

ANA_FRO12M_CLK_ENAW
Values that can be written to the field ANA_FRO12M_CLK_ENA

CLKIN_ENAR
Possible values of the field CLKIN_ENA

CLKIN_ENAW
Values that can be written to the field CLKIN_ENA

FLASH48MHZ_ENAR
Possible values of the field FLASH48MHZ_ENA

FLASH48MHZ_ENAW
Values that can be written to the field FLASH48MHZ_ENA

FRO1MHZ_UTICK_ENAR
Possible values of the field FRO1MHZ_UTICK_ENA

FRO1MHZ_CLK_ENAR
Possible values of the field FRO1MHZ_CLK_ENA

FRO1MHZ_UTICK_ENAW
Values that can be written to the field FRO1MHZ_UTICK_ENA

FRO1MHZ_CLK_ENAW
Values that can be written to the field FRO1MHZ_CLK_ENA

FRO12MHZ_FREQM_ENAR
Possible values of the field FRO12MHZ_FREQM_ENA

FRO12MHZ_FREQM_ENAW
Values that can be written to the field FRO12MHZ_FREQM_ENA

FRO_HF_FREQM_ENAR
Possible values of the field FRO_HF_FREQM_ENA

FRO_HF_FREQM_ENAW
Values that can be written to the field FRO_HF_FREQM_ENA

PLU_DEGLITCH_CLK_ENAR
Possible values of the field PLU_DEGLITCH_CLK_ENA

PLU_DEGLITCH_CLK_ENAW
Values that can be written to the field PLU_DEGLITCH_CLK_ENA

XO_CAL_CLK_ENAR
Possible values of the field XO_CAL_CLK_ENA

XO_CAL_CLK_ENAW
Values that can be written to the field XO_CAL_CLK_ENA

XTAL32MHZ_FREQM_ENAR
Possible values of the field XTAL32MHZ_FREQM_ENA

XTAL32MHZ_FREQM_ENAW
Values that can be written to the field XTAL32MHZ_FREQM_ENA