Module lp55s69_pac::syscon::presetctrl1

Peripheral reset control 1

Structs

R
Value read from the register

W
Value to write to the register

_EZHA_RSTW
Proxy

_EZHB_RSTW
Proxy

_FC0_RSTW
Proxy

_FC1_RSTW
Proxy

_FC2_RSTW
Proxy

_FC3_RSTW
Proxy

_FC4_RSTW
Proxy

_FC5_RSTW
Proxy

_FC6_RSTW
Proxy

_FC7_RSTW
Proxy

_MRT_RSTW
Proxy

_OSTIMER0_RSTW
Proxy

_PVT_RSTW
Proxy

_SCT0_RSTW
Proxy

_SCTIPU_RSTW
Proxy

_TIMER0_RSTW
Proxy

_TIMER1_RSTW
Proxy

_TIMER2_RSTW
Proxy

_USB0_DEV_RSTW
Proxy

_UTICK0_RSTW
Proxy

Enums

EZHA_RSTR
Possible values of the field EZHA_RST

EZHA_RSTW
Values that can be written to the field EZHA_RST

EZHB_RSTR
Possible values of the field EZHB_RST

EZHB_RSTW
Values that can be written to the field EZHB_RST

FC0_RSTR
Possible values of the field FC0_RST

FC0_RSTW
Values that can be written to the field FC0_RST

FC1_RSTR
Possible values of the field FC1_RST

FC1_RSTW
Values that can be written to the field FC1_RST

FC2_RSTR
Possible values of the field FC2_RST

FC2_RSTW
Values that can be written to the field FC2_RST

FC3_RSTR
Possible values of the field FC3_RST

FC3_RSTW
Values that can be written to the field FC3_RST

FC4_RSTR
Possible values of the field FC4_RST

FC4_RSTW
Values that can be written to the field FC4_RST

FC5_RSTR
Possible values of the field FC5_RST

FC5_RSTW
Values that can be written to the field FC5_RST

FC6_RSTR
Possible values of the field FC6_RST

FC6_RSTW
Values that can be written to the field FC6_RST

FC7_RSTR
Possible values of the field FC7_RST

FC7_RSTW
Values that can be written to the field FC7_RST

MRT_RSTR
Possible values of the field MRT_RST

MRT_RSTW
Values that can be written to the field MRT_RST

OSTIMER0_RSTR
Possible values of the field OSTIMER0_RST

OSTIMER0_RSTW
Values that can be written to the field OSTIMER0_RST

PVT_RSTR
Possible values of the field PVT_RST

PVT_RSTW
Values that can be written to the field PVT_RST

SCT0_RSTR
Possible values of the field SCT0_RST

SCT0_RSTW
Values that can be written to the field SCT0_RST

SCTIPU_RSTR
Possible values of the field SCTIPU_RST

SCTIPU_RSTW
Values that can be written to the field SCTIPU_RST

TIMER0_RSTR
Possible values of the field TIMER0_RST

TIMER0_RSTW
Values that can be written to the field TIMER0_RST

TIMER1_RSTR
Possible values of the field TIMER1_RST

TIMER1_RSTW
Values that can be written to the field TIMER1_RST

TIMER2_RSTR
Possible values of the field TIMER2_RST

TIMER2_RSTW
Values that can be written to the field TIMER2_RST

USB0_DEV_RSTR
Possible values of the field USB0_DEV_RST

USB0_DEV_RSTW
Values that can be written to the field USB0_DEV_RST

UTICK0_RSTR
Possible values of the field UTICK0_RST

UTICK0_RSTW
Values that can be written to the field UTICK0_RST