Module lp55s69_pac::syscon::starter0

Start logic wake-up enable register

Structs

R
Value read from the register

W
Value to write to the register

_ADC0W
Proxy

_ADC0_THCMP_OVRW
Proxy

_CTIMER0W
Proxy

_CTIMER1W
Proxy

_CTIMER3W
Proxy

_EZH_ARCH_B0W
Proxy

_FLEXINT0W
Proxy

_FLEXINT1W
Proxy

_FLEXINT2W
Proxy

_FLEXINT3W
Proxy

_FLEXINT4W
Proxy

_FLEXINT5W
Proxy

_FLEXINT6W
Proxy

_FLEXINT7W
Proxy

_GINT0W
Proxy

_GINT1W
Proxy

_MRT0W
Proxy

_PIO_INT0W
Proxy

_PIO_INT1W
Proxy

_PIO_INT2W
Proxy

_PIO_INT3W
Proxy

_RTC_LITE0W
Proxy

_SCT0W
Proxy

_SDMA0W
Proxy

_SYSW
Proxy

_USB0W
Proxy

_USB0_NEEDCLKW
Proxy

_UTICK0W
Proxy

_WAKEUP_MAILBOX0W
Proxy

Enums

ADC0R
Possible values of the field ADC0

ADC0W
Values that can be written to the field ADC0

ADC0_THCMP_OVRR
Possible values of the field ADC0_THCMP_OVR

ADC0_THCMP_OVRW
Values that can be written to the field ADC0_THCMP_OVR

CTIMER0R
Possible values of the field CTIMER0

CTIMER0W
Values that can be written to the field CTIMER0

CTIMER1R
Possible values of the field CTIMER1

CTIMER1W
Values that can be written to the field CTIMER1

CTIMER3R
Possible values of the field CTIMER3

CTIMER3W
Values that can be written to the field CTIMER3

EZH_ARCH_B0R
Possible values of the field EZH_ARCH_B0

EZH_ARCH_B0W
Values that can be written to the field EZH_ARCH_B0

FLEXINT0R
Possible values of the field FLEXINT0

FLEXINT0W
Values that can be written to the field FLEXINT0

FLEXINT1R
Possible values of the field FLEXINT1

FLEXINT1W
Values that can be written to the field FLEXINT1

FLEXINT2R
Possible values of the field FLEXINT2

FLEXINT2W
Values that can be written to the field FLEXINT2

FLEXINT3R
Possible values of the field FLEXINT3

FLEXINT3W
Values that can be written to the field FLEXINT3

FLEXINT4R
Possible values of the field FLEXINT4

FLEXINT4W
Values that can be written to the field FLEXINT4

FLEXINT5R
Possible values of the field FLEXINT5

FLEXINT5W
Values that can be written to the field FLEXINT5

FLEXINT6R
Possible values of the field FLEXINT6

FLEXINT6W
Values that can be written to the field FLEXINT6

FLEXINT7R
Possible values of the field FLEXINT7

FLEXINT7W
Values that can be written to the field FLEXINT7

GINT0R
Possible values of the field GINT0

GINT0W
Values that can be written to the field GINT0

GINT1R
Possible values of the field GINT1

GINT1W
Values that can be written to the field GINT1

MRT0R
Possible values of the field MRT0

MRT0W
Values that can be written to the field MRT0

PIO_INT0R
Possible values of the field PIO_INT0

PIO_INT0W
Values that can be written to the field PIO_INT0

PIO_INT1R
Possible values of the field PIO_INT1

PIO_INT1W
Values that can be written to the field PIO_INT1

PIO_INT2R
Possible values of the field PIO_INT2

PIO_INT2W
Values that can be written to the field PIO_INT2

PIO_INT3R
Possible values of the field PIO_INT3

PIO_INT3W
Values that can be written to the field PIO_INT3

RTC_LITE0R
Possible values of the field RTC_LITE0

RTC_LITE0W
Values that can be written to the field RTC_LITE0

SCT0R
Possible values of the field SCT0

SCT0W
Values that can be written to the field SCT0

SDMA0R
Possible values of the field SDMA0

SDMA0W
Values that can be written to the field SDMA0

SYSR
Possible values of the field SYS

SYSW
Values that can be written to the field SYS

USB0R
Possible values of the field USB0

USB0W
Values that can be written to the field USB0

USB0_NEEDCLKR
Possible values of the field USB0_NEEDCLK

USB0_NEEDCLKW
Values that can be written to the field USB0_NEEDCLK

UTICK0R
Possible values of the field UTICK0

UTICK0W
Values that can be written to the field UTICK0

WAKEUP_MAILBOX0R
Possible values of the field WAKEUP_MAILBOX0

WAKEUP_MAILBOX0W
Values that can be written to the field WAKEUP_MAILBOX0