Struct lp55s69_pac::syscon::starterclr1::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn gpio_int04_clr(&mut self) -> _GPIO_INT04_CLRW

Bit 0 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn gpio_int05_clr(&mut self) -> _GPIO_INT05_CLRW

Bit 1 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn gpio_int06_clr(&mut self) -> _GPIO_INT06_CLRW

Bit 2 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn gpio_int07_clr(&mut self) -> _GPIO_INT07_CLRW

Bit 3 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn ctimer2_clr(&mut self) -> _CTIMER2_CLRW

Bit 4 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn ctimer4_clr(&mut self) -> _CTIMER4_CLRW

Bit 5 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn os_event_clr(&mut self) -> _OS_EVENT_CLRW

Bit 6 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sdio_clr(&mut self) -> _SDIO_CLRW

Bit 10 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn usb1_clr(&mut self) -> _USB1_CLRW

Bit 15 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn usb1_needclk_clr(&mut self) -> _USB1_NEEDCLK_CLRW

Bit 16 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sec_hypervisor_call_clr(&mut self) -> _SEC_HYPERVISOR_CALL_CLRW

Bit 17 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sec_gpio_int00_clr(&mut self) -> _SEC_GPIO_INT00_CLRW

Bit 18 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sec_gpio_int01_clr(&mut self) -> _SEC_GPIO_INT01_CLRW

Bit 19 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn plu_clr(&mut self) -> _PLU_CLRW

Bit 20 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sec_vio_clr(&mut self) -> _SEC_VIO_CLRW

Bit 21 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sha_clr(&mut self) -> _SHA_CLRW

Bit 22 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn caser_clr(&mut self) -> _CASER_CLRW

Bit 23 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn qddkey_clr(&mut self) -> _QDDKEY_CLRW

Bit 24 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn pq_clr(&mut self) -> _PQ_CLRW

Bit 25 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn sdma1_clr(&mut self) -> _SDMA1_CLRW

Bit 26 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn lspi_hs_clr(&mut self) -> _LSPI_HS_CLRW

Bit 27 - Writing ones to this register clears the corresponding bit in the STARTER1 register.

pub fn wakeuppads_clr(&mut self) -> _WAKEUPPADS_CLRW

Bit 31 - Writing ones to this register clears the corresponding bit in the STARTER1 register.