Struct lp55s69_pac::syscon::starterset0::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn sys_set(&mut self) -> _SYS_SETW

Bit 0 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn sdma0_set(&mut self) -> _SDMA0_SETW

Bit 1 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_globalint0_set(&mut self) -> _GPIO_GLOBALINT0_SETW

Bit 2 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_globalint1_set(&mut self) -> _GPIO_GLOBALINT1_SETW

Bit 3 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_int00_set(&mut self) -> _GPIO_INT00_SETW

Bit 4 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_int01_set(&mut self) -> _GPIO_INT01_SETW

Bit 5 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_int02_set(&mut self) -> _GPIO_INT02_SETW

Bit 6 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn gpio_int03_set(&mut self) -> _GPIO_INT03_SETW

Bit 7 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn utick0_set(&mut self) -> _UTICK0_SETW

Bit 8 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn mrt0_set(&mut self) -> _MRT0_SETW

Bit 9 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn ctimer0_set(&mut self) -> _CTIMER0_SETW

Bit 10 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn ctimer1_set(&mut self) -> _CTIMER1_SETW

Bit 11 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn sct0_set(&mut self) -> _SCT0_SETW

Bit 12 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn ctimer3_set(&mut self) -> _CTIMER3_SETW

Bit 13 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint0_set(&mut self) -> _FLEXINT0_SETW

Bit 14 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint1_set(&mut self) -> _FLEXINT1_SETW

Bit 15 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint2_set(&mut self) -> _FLEXINT2_SETW

Bit 16 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint3_set(&mut self) -> _FLEXINT3_SETW

Bit 17 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint4_set(&mut self) -> _FLEXINT4_SETW

Bit 18 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint5_set(&mut self) -> _FLEXINT5_SETW

Bit 19 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint6_set(&mut self) -> _FLEXINT6_SETW

Bit 20 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn flexint7_set(&mut self) -> _FLEXINT7_SETW

Bit 21 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn adc0_set(&mut self) -> _ADC0_SETW

Bit 22 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn adc0_thcmp_ovr_set(&mut self) -> _ADC0_THCMP_OVR_SETW

Bit 24 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn usb0_needclk_set(&mut self) -> _USB0_NEEDCLK_SETW

Bit 27 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn usb0_set(&mut self) -> _USB0_SETW

Bit 28 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn rtc_lite0_set(&mut self) -> _RTC_LITE0_SETW

Bit 29 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn ezh_arch_b0_set(&mut self) -> _EZH_ARCH_B0_SETW

Bit 30 - Writing ones to this register sets the corresponding bit in the STARTER0 register.

pub fn wakeup_mailbox0_set(&mut self) -> _WAKEUP_MAILBOX0_SETW

Bit 31 - Writing ones to this register sets the corresponding bit in the STARTER0 register.