Struct lp55s69_pac::syscon::starterset1::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn gpio_int04_set(&mut self) -> _GPIO_INT04_SETW

Bit 0 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn gpio_int05_set(&mut self) -> _GPIO_INT05_SETW

Bit 1 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn gpio_int06_set(&mut self) -> _GPIO_INT06_SETW

Bit 2 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn gpio_int07_set(&mut self) -> _GPIO_INT07_SETW

Bit 3 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn ctimer2_set(&mut self) -> _CTIMER2_SETW

Bit 4 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn ctimer4_set(&mut self) -> _CTIMER4_SETW

Bit 5 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn os_event_set(&mut self) -> _OS_EVENT_SETW

Bit 6 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sdio_set(&mut self) -> _SDIO_SETW

Bit 10 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn usb1_set(&mut self) -> _USB1_SETW

Bit 15 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn usb1_needclk_set(&mut self) -> _USB1_NEEDCLK_SETW

Bit 16 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sec_hypervisor_call_set(&mut self) -> _SEC_HYPERVISOR_CALL_SETW

Bit 17 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sec_gpio_int00_set(&mut self) -> _SEC_GPIO_INT00_SETW

Bit 18 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sec_gpio_int01_set(&mut self) -> _SEC_GPIO_INT01_SETW

Bit 19 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn plu_set(&mut self) -> _PLU_SETW

Bit 20 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sec_vio_set(&mut self) -> _SEC_VIO_SETW

Bit 21 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sha_set(&mut self) -> _SHA_SETW

Bit 22 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn caser_set(&mut self) -> _CASER_SETW

Bit 23 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn qddkey_set(&mut self) -> _QDDKEY_SETW

Bit 24 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn pq_set(&mut self) -> _PQ_SETW

Bit 25 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn sdma1_set(&mut self) -> _SDMA1_SETW

Bit 26 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn lspi_hs_set(&mut self) -> _LSPI_HS_SETW

Bit 27 - Writing ones to this register sets the corresponding bit in the STARTER1 register.

pub fn wakeuppads_set(&mut self) -> _WAKEUPPADS_SETW

Bit 31 - Writing ones to this register sets the corresponding bit in the STARTER1 register.