Module lp55s69_pac::sysctl::sharedctrlset

Selects sources and data combinations for shared signal set 0.

Structs

R
Value read from the register

W
Value to write to the register

_F20DATAOUTENW
Proxy

_FC0DATAOUTENW
Proxy

_FC1DATAOUTENW
Proxy

_FC3DATAOUTENW
Proxy

_FC4DATAOUTENW
Proxy

_FC5DATAOUTENW
Proxy

_FC6DATAOUTENW
Proxy

_FC7DATAOUTENW
Proxy

_SHAREDDATASELW
Proxy

_SHAREDSCKSELW
Proxy

_SHAREDWSSELW
Proxy

Enums

F20DATAOUTENR
Possible values of the field F20DATAOUTEN

F20DATAOUTENW
Values that can be written to the field F20DATAOUTEN

FC0DATAOUTENR
Possible values of the field FC0DATAOUTEN

FC0DATAOUTENW
Values that can be written to the field FC0DATAOUTEN

FC1DATAOUTENR
Possible values of the field FC1DATAOUTEN

FC1DATAOUTENW
Values that can be written to the field FC1DATAOUTEN

FC3DATAOUTENR
Possible values of the field FC3DATAOUTEN

FC3DATAOUTENW
Values that can be written to the field FC3DATAOUTEN

FC4DATAOUTENR
Possible values of the field FC4DATAOUTEN

FC4DATAOUTENW
Values that can be written to the field FC4DATAOUTEN

FC5DATAOUTENR
Possible values of the field FC5DATAOUTEN

FC5DATAOUTENW
Values that can be written to the field FC5DATAOUTEN

FC6DATAOUTENR
Possible values of the field FC6DATAOUTEN

FC6DATAOUTENW
Values that can be written to the field FC6DATAOUTEN

FC7DATAOUTENR
Possible values of the field FC7DATAOUTEN

FC7DATAOUTENW
Values that can be written to the field FC7DATAOUTEN

SHAREDDATASELR
Possible values of the field SHAREDDATASEL

SHAREDDATASELW
Values that can be written to the field SHAREDDATASEL

SHAREDSCKSELR
Possible values of the field SHAREDSCKSEL

SHAREDSCKSELW
Values that can be written to the field SHAREDSCKSEL

SHAREDWSSELR
Possible values of the field SHAREDWSSEL

SHAREDWSSELW
Values that can be written to the field SHAREDWSSEL