Struct lp55s69_pac::usart0::cfg::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn enable(&self) -> ENABLER

Bit 0 - USART Enable.

pub fn datalen(&self) -> DATALENR

Bits 2:3 - Selects the data size for the USART.

pub fn paritysel(&self) -> PARITYSELR

Bits 4:5 - Selects what type of parity is used by the USART.

pub fn stoplen(&self) -> STOPLENR

Bit 6 - Number of stop bits appended to transmitted data. Only a single stop bit is required for received data.

pub fn mode32k(&self) -> MODE32KR

Bit 7 - Selects standard or 32 kHz clocking mode.

pub fn linmode(&self) -> LINMODER

Bit 8 - LIN break mode enable.

pub fn ctsen(&self) -> CTSENR

Bit 9 - CTS Enable. Determines whether CTS is used for flow control. CTS can be from the input pin, or from the USART's own RTS if loopback mode is enabled.

pub fn syncen(&self) -> SYNCENR

Bit 11 - Selects synchronous or asynchronous operation.

pub fn clkpol(&self) -> CLKPOLR

Bit 12 - Selects the clock polarity and sampling edge of received data in synchronous mode.

pub fn syncmst(&self) -> SYNCMSTR

Bit 14 - Synchronous mode Master select.

pub fn loop_(&self) -> LOOPR

Bit 15 - Selects data loopback mode.

pub fn oeta(&self) -> OETAR

Bit 18 - Output Enable Turnaround time enable for RS-485 operation.

pub fn autoaddr(&self) -> AUTOADDRR

Bit 19 - Automatic Address matching enable.

pub fn oesel(&self) -> OESELR

Bit 20 - Output Enable Select.

pub fn oepol(&self) -> OEPOLR

Bit 21 - Output Enable Polarity.

pub fn rxpol(&self) -> RXPOLR

Bit 22 - Receive data polarity.

pub fn txpol(&self) -> TXPOLR

Bit 23 - Transmit data polarity.