Module lp55s69_pac::usb0::devcmdstat

USB Device Command/Status register

Structs

DCONR
Value of the field

DCON_CR
Value of the field

DEV_ADDRR
Value of the field

DEV_ENR
Value of the field

DRES_CR
Value of the field

DSUSR
Value of the field

DSUS_CR
Value of the field

LPM_REWPR
Value of the field

LPM_SUSR
Value of the field

R
Value read from the register

SETUPR
Value of the field

VBUSDEBOUNCEDR
Value of the field

W
Value to write to the register

_DCONW
Proxy

_DCON_CW
Proxy

_DEV_ADDRW
Proxy

_DEV_ENW
Proxy

_DRES_CW
Proxy

_DSUSW
Proxy

_DSUS_CW
Proxy

_FORCE_NEEDCLKW
Proxy

_INTONNAK_AIW
Proxy

_INTONNAK_AOW
Proxy

_INTONNAK_CIW
Proxy

_INTONNAK_COW
Proxy

_LPM_SUPW
Proxy

_LPM_SUSW
Proxy

_SETUPW
Proxy

Enums

FORCE_NEEDCLKR
Possible values of the field FORCE_NEEDCLK

FORCE_NEEDCLKW
Values that can be written to the field FORCE_NEEDCLK

INTONNAK_AIR
Possible values of the field INTONNAK_AI

INTONNAK_AIW
Values that can be written to the field INTONNAK_AI

INTONNAK_AOR
Possible values of the field INTONNAK_AO

INTONNAK_AOW
Values that can be written to the field INTONNAK_AO

INTONNAK_CIR
Possible values of the field INTONNAK_CI

INTONNAK_CIW
Values that can be written to the field INTONNAK_CI

INTONNAK_COR
Possible values of the field INTONNAK_CO

INTONNAK_COW
Values that can be written to the field INTONNAK_CO

LPM_SUPR
Possible values of the field LPM_SUP

LPM_SUPW
Values that can be written to the field LPM_SUP