Struct lp55s69_pac::usb0::devcmdstat::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn dev_addr(&self) -> DEV_ADDRR

Bits 0:6 - USB device address. After bus reset, the address is reset to 0x00. If the enable bit is set, the device will respond on packets for function address DEV_ADDR. When receiving a SetAddress Control Request from the USB host, software must program the new address before completing the status phase of the SetAddress Control Request.

pub fn dev_en(&self) -> DEV_ENR

Bit 7 - USB device enable. If this bit is set, the HW will start responding on packets for function address DEV_ADDR.

pub fn setup(&self) -> SETUPR

Bit 8 - SETUP token received. If a SETUP token is received and acknowledged by the device, this bit is set. As long as this bit is set all received IN and OUT tokens will be NAKed by HW. SW must clear this bit by writing a one. If this bit is zero, HW will handle the tokens to the CTRL EP0 as indicated by the CTRL EP0 IN and OUT data information programmed by SW.

pub fn force_needclk(&self) -> FORCE_NEEDCLKR

Bit 9 - Forces the NEEDCLK output to always be on:

pub fn lpm_sup(&self) -> LPM_SUPR

Bit 11 - LPM Supported:

pub fn intonnak_ao(&self) -> INTONNAK_AOR

Bit 12 - Interrupt on NAK for interrupt and bulk OUT EP

pub fn intonnak_ai(&self) -> INTONNAK_AIR

Bit 13 - Interrupt on NAK for interrupt and bulk IN EP

pub fn intonnak_co(&self) -> INTONNAK_COR

Bit 14 - Interrupt on NAK for control OUT EP

pub fn intonnak_ci(&self) -> INTONNAK_CIR

Bit 15 - Interrupt on NAK for control IN EP

pub fn dcon(&self) -> DCONR

Bit 16 - Device status - connect. The connect bit must be set by SW to indicate that the device must signal a connect. The pull-up resistor on USB_DP will be enabled when this bit is set and the VBUSDEBOUNCED bit is one.

pub fn dsus(&self) -> DSUSR

Bit 17 - Device status - suspend. The suspend bit indicates the current suspend state. It is set to 1 when the device hasn't seen any activity on its upstream port for more than 3 milliseconds. It is reset to 0 on any activity. When the device is suspended (Suspend bit DSUS = 1) and the software writes a 0 to it, the device will generate a remote wake-up. This will only happen when the device is connected (Connect bit = 1). When the device is not connected or not suspended, a writing a 0 has no effect. Writing a 1 never has an effect.

pub fn lpm_sus(&self) -> LPM_SUSR

Bit 19 - Device status - LPM Suspend. This bit represents the current LPM suspend state. It is set to 1 by HW when the device has acknowledged the LPM request from the USB host and the Token Retry Time of 10 ms has elapsed. When the device is in the LPM suspended state (LPM suspend bit = 1) and the software writes a zero to this bit, the device will generate a remote walk-up. Software can only write a zero to this bit when the LPM_REWP bit is set to 1. HW resets this bit when it receives a host initiated resume. HW only updates the LPM_SUS bit when the LPM_SUPP bit is equal to one.

pub fn lpm_rewp(&self) -> LPM_REWPR

Bit 20 - LPM Remote Wake-up Enabled by USB host. HW sets this bit to one when the bRemoteWake bit in the LPM extended token is set to 1. HW will reset this bit to 0 when it receives the host initiated LPM resume, when a remote wake-up is sent by the device or when a USB bus reset is received. Software can use this bit to check if the remote wake-up feature is enabled by the host for the LPM transaction.

pub fn dcon_c(&self) -> DCON_CR

Bit 24 - Device status - connect change. The Connect Change bit is set when the device's pull-up resistor is disconnected because VBus disappeared. The bit is reset by writing a one to it.

pub fn dsus_c(&self) -> DSUS_CR

Bit 25 - Device status - suspend change. The suspend change bit is set to 1 when the suspend bit toggles. The suspend bit can toggle because: - The device goes in the suspended state - The device is disconnected - The device receives resume signaling on its upstream port. The bit is reset by writing a one to it.

pub fn dres_c(&self) -> DRES_CR

Bit 26 - Device status - reset change. This bit is set when the device received a bus reset. On a bus reset the device will automatically go to the default state (unconfigured and responding to address 0). The bit is reset by writing a one to it.

pub fn vbusdebounced(&self) -> VBUSDEBOUNCEDR

Bit 28 - This bit indicates if Vbus is detected or not. The bit raises immediately when Vbus becomes high. It drops to zero if Vbus is low for at least 3 ms. If this bit is high and the DCon bit is set, the HW will enable the pull-up resistor to signal a connect.