Enum lp55s69_pac::usb0::info::ERR_CODER

pub enum ERR_CODER {
    NO_ERROR,
    PID_ENCODING_ERROR,
    PID_UNKNOWN,
    PACKET_UNEXPECTED,
    TOKEN_CRC_ERROR,
    DATA_CRC_ERROR,
    TIMEOUT,
    BABBLE,
    TRUNCATED_EOP,
    SENT_RECEIVED_NAK,
    SENT_STALL,
    OVERRUN,
    SENT_EMPTY_PACKET,
    BITSTUFF_ERROR,
    SYNC_ERROR,
    WRONG_DATA_TOGGLE,
}

Possible values of the field ERR_CODE

Variants
NO_ERROR

No error

PID_ENCODING_ERROR

PID encoding error

PID_UNKNOWN

PID unknown

PACKET_UNEXPECTED

Packet unexpected

TOKEN_CRC_ERROR

Token CRC error

DATA_CRC_ERROR

Data CRC error

TIMEOUT

Time out

BABBLE

Babble

TRUNCATED_EOP

Truncated EOP

SENT_RECEIVED_NAK

Sent/Received NAK

SENT_STALL

Sent Stall

OVERRUN

Overrun

SENT_EMPTY_PACKET

Sent empty packet

BITSTUFF_ERROR

Bitstuff error

SYNC_ERROR

Sync error

WRONG_DATA_TOGGLE

Wrong data toggle

Methods

impl ERR_CODER

pub fn bits(&self) -> u8

Value of the field as raw bits

pub fn is_no_error(&self) -> bool

Checks if the value of the field is NO_ERROR

pub fn is_pid_encoding_error(&self) -> bool

Checks if the value of the field is PID_ENCODING_ERROR

pub fn is_pid_unknown(&self) -> bool

Checks if the value of the field is PID_UNKNOWN

pub fn is_packet_unexpected(&self) -> bool

Checks if the value of the field is PACKET_UNEXPECTED

pub fn is_token_crc_error(&self) -> bool

Checks if the value of the field is TOKEN_CRC_ERROR

pub fn is_data_crc_error(&self) -> bool

Checks if the value of the field is DATA_CRC_ERROR

pub fn is_timeout(&self) -> bool

Checks if the value of the field is TIMEOUT

pub fn is_babble(&self) -> bool

Checks if the value of the field is BABBLE

pub fn is_truncated_eop(&self) -> bool

Checks if the value of the field is TRUNCATED_EOP

pub fn is_sent_received_nak(&self) -> bool

Checks if the value of the field is SENT_RECEIVED_NAK

pub fn is_sent_stall(&self) -> bool

Checks if the value of the field is SENT_STALL

pub fn is_overrun(&self) -> bool

Checks if the value of the field is OVERRUN

pub fn is_sent_empty_packet(&self) -> bool

Checks if the value of the field is SENT_EMPTY_PACKET

pub fn is_bitstuff_error(&self) -> bool

Checks if the value of the field is BITSTUFF_ERROR

pub fn is_sync_error(&self) -> bool

Checks if the value of the field is SYNC_ERROR

pub fn is_wrong_data_toggle(&self) -> bool

Checks if the value of the field is WRONG_DATA_TOGGLE