Struct lp55s69_pac::usbhsd::devcmdstat::R

pub struct R { /* fields omitted */ }

Value read from the register

Methods

impl R

pub fn bits(&self) -> u32

Value of the register as raw bits

pub fn dev_addr(&self) -> DEV_ADDRR

Bits 0:6 - USB device address.

pub fn dev_en(&self) -> DEV_ENR

Bit 7 - USB device enable.

pub fn setup(&self) -> SETUPR

Bit 8 - SETUP token received.

pub fn force_needclk(&self) -> FORCE_NEEDCLKR

Bit 9 - Forces the NEEDCLK output to always be on:.

pub fn force_vbus(&self) -> FORCE_VBUSR

Bit 10 - If this bit is set to 1, the VBUS voltage indicators from the PHY are overruled.

pub fn lpm_sup(&self) -> LPM_SUPR

Bit 11 - LPM Supported:.

pub fn intonnak_ao(&self) -> INTONNAK_AOR

Bit 12 - Interrupt on NAK for interrupt and bulk OUT EP:.

pub fn intonnak_ai(&self) -> INTONNAK_AIR

Bit 13 - Interrupt on NAK for interrupt and bulk IN EP:.

pub fn intonnak_co(&self) -> INTONNAK_COR

Bit 14 - Interrupt on NAK for control OUT EP:.

pub fn intonnak_ci(&self) -> INTONNAK_CIR

Bit 15 - Interrupt on NAK for control IN EP:.

pub fn dcon(&self) -> DCONR

Bit 16 - Device status - connect.

pub fn dsus(&self) -> DSUSR

Bit 17 - Device status - suspend.

pub fn lpm_sus(&self) -> LPM_SUSR

Bit 19 - Device status - LPM Suspend.

pub fn lpm_rewp(&self) -> LPM_REWPR

Bit 20 - LPM Remote Wake-up Enabled by USB host.

pub fn speed(&self) -> SPEEDR

Bits 22:23 - This field indicates the speed at which the device operates: 00b: reserved 01b: full-speed 10b: high-speed 11b: super-speed (reserved for future use).

pub fn dcon_c(&self) -> DCON_CR

Bit 24 - Device status - connect change.

pub fn dsus_c(&self) -> DSUS_CR

Bit 25 - Device status - suspend change.

pub fn dres_c(&self) -> DRES_CR

Bit 26 - Device status - reset change.

pub fn vbus_debounced(&self) -> VBUS_DEBOUNCEDR

Bit 28 - This bit indicates if VBUS is detected or not.

pub fn phy_test_mode(&self) -> PHY_TEST_MODER

Bits 29:31 - This field is written by firmware to put the PHY into a test mode as defined by the USB2.