Struct lp55s69_pac::usbhsh::RegisterBlock

#[repr(C)]

pub struct RegisterBlock {
    pub caplength_chipid: CAPLENGTH_CHIPID,
    pub hcsparams: HCSPARAMS,
    pub hccparams: HCCPARAMS,
    pub fladj_frindex: FLADJ_FRINDEX,
    pub atl_ptd_base_addr: ATL_PTD_BASE_ADDR,
    pub iso_ptd_base_addr: ISO_PTD_BASE_ADDR,
    pub int_ptd_base_addr: INT_PTD_BASE_ADDR,
    pub data_payload_base_addr: DATA_PAYLOAD_BASE_ADDR,
    pub usbcmd: USBCMD,
    pub usbsts: USBSTS,
    pub usbintr: USBINTR,
    pub portsc1: PORTSC1,
    pub atl_ptd_done_map: ATL_PTD_DONE_MAP,
    pub atl_ptd_skip_map: ATL_PTD_SKIP_MAP,
    pub iso_ptd_done_map: ISO_PTD_DONE_MAP,
    pub iso_ptd_skip_map: ISO_PTD_SKIP_MAP,
    pub int_ptd_done_map: INT_PTD_DONE_MAP,
    pub int_ptd_skip_map: INT_PTD_SKIP_MAP,
    pub last_ptd_inuse: LAST_PTD_INUSE,
    pub utmiplus_ulpi_debug: UTMIPLUS_ULPI_DEBUG,
    pub portmode: PORTMODE,
}

Register block

Fields
caplength_chipid: CAPLENGTH_CHIPID

0x00 - This register contains the offset value towards the start of the operational register space and the version number of the IP block

hcsparams: HCSPARAMS

0x04 - Host Controller Structural Parameters

hccparams: HCCPARAMS

0x08 - Host Controller Capability Parameters

fladj_frindex: FLADJ_FRINDEX

0x0c - Frame Length Adjustment

atl_ptd_base_addr: ATL_PTD_BASE_ADDR

0x10 - Memory base address where ATL PTD0 is stored

iso_ptd_base_addr: ISO_PTD_BASE_ADDR

0x14 - Memory base address where ISO PTD0 is stored

int_ptd_base_addr: INT_PTD_BASE_ADDR

0x18 - Memory base address where INT PTD0 is stored

data_payload_base_addr: DATA_PAYLOAD_BASE_ADDR

0x1c - Memory base address that indicates the start of the data payload buffers

usbcmd: USBCMD

0x20 - USB Command register

usbsts: USBSTS

0x24 - USB Interrupt Status register

usbintr: USBINTR

0x28 - USB Interrupt Enable register

portsc1: PORTSC1

0x2c - Port Status and Control register

atl_ptd_done_map: ATL_PTD_DONE_MAP

0x30 - Done map for each ATL PTD

atl_ptd_skip_map: ATL_PTD_SKIP_MAP

0x34 - Skip map for each ATL PTD

iso_ptd_done_map: ISO_PTD_DONE_MAP

0x38 - Done map for each ISO PTD

iso_ptd_skip_map: ISO_PTD_SKIP_MAP

0x3c - Skip map for each ISO PTD

int_ptd_done_map: INT_PTD_DONE_MAP

0x40 - Done map for each INT PTD

int_ptd_skip_map: INT_PTD_SKIP_MAP

0x44 - Skip map for each INT PTD

last_ptd_inuse: LAST_PTD_INUSE

0x48 - Marks the last PTD in the list for ISO, INT and ATL

utmiplus_ulpi_debug: UTMIPLUS_ULPI_DEBUG

0x4c - Register to read/write registers in the attached USB PHY

portmode: PORTMODE

0x50 - Controls the port if it is attached to the host block or the device block