Struct lp55s69_pac::usbphy::debug0_clr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn otgidpiolock(&mut self) -> _OTGIDPIOLOCKW

Bit 0 - Once OTG ID from STATUS_OTGID_STATUS is sampled, use this to hold the value

pub fn debug_interface_hold(&mut self) -> _DEBUG_INTERFACE_HOLDW

Bit 1 - Use holding registers to assist in timing for external UTMI interface.

pub fn hstpulldown(&mut self) -> _HSTPULLDOWNW

Bits 2:3 - This bit field selects whether to connect pulldown resistors on the USB_DP/USB_DM pins if the corresponding pulldown overdrive mode is enabled through DEBUG[5:4} Set bit 3 to value 1'b1 to connect the 15ohm pulldown on USB_DP line

pub fn enhstpulldown(&mut self) -> _ENHSTPULLDOWNW

Bits 4:5 - This bit field selects host pulldown overdrive mode

pub fn tx2rxcount(&mut self) -> _TX2RXCOUNTW

Bits 8:11 - Delay in between the end of transmit to the beginning of receive

pub fn entx2rxcount(&mut self) -> _ENTX2RXCOUNTW

Bit 12 - Set this bit to allow a countdown to transition in between TX and RX.

pub fn squelchresetcount(&mut self) -> _SQUELCHRESETCOUNTW

Bits 16:20 - Delay in between the detection of squelch to the reset of high-speed RX.

pub fn ensquelchreset(&mut self) -> _ENSQUELCHRESETW

Bit 24 - Set bit to allow squelch to reset high-speed receive.

pub fn squelchresetlength(&mut self) -> _SQUELCHRESETLENGTHW

Bits 25:28 - Duration of RESET in terms of the number of 480-MHz cycles.

pub fn host_resume_debug(&mut self) -> _HOST_RESUME_DEBUGW

Bit 29 - Choose to trigger the host resume SE0 with HOST_FORCE_LS_SE0 = 0 or UTMI_SUSPEND = 1.

pub fn clkgate(&mut self) -> _CLKGATEW

Bit 30 - Gate Test Clocks